Refactor partition services
- Introduce PartitionStateManager
- Introduce MigrationManager
- Introduce ReplicaSafetyChecker
- Pull out MigrationThread
- Pull out PublishPartitionRuntimeStateTask
- Pull out InternalPartitionListener
- Pull out PartitionEventManager
- Prepare migration tasks after a node is removed
	* Null partition replicas which are assigned to removed node
	* Swap 1st and 2nd replica if it was the owner
	* Migrate null-ed replica to the last non-null replica owner in the list
