+++ hazelcast/src/main/java/com/hazelcast/internal/partition/impl/InternalMigrationListener.java
@@ -12,10 +12,9 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *
  */
 
-package com.hazelcast.partition.impl;
+package com.hazelcast.internal.partition.impl;
 
 
 import com.hazelcast.internal.partition.MigrationInfo;
