+++ hazelcast/src/main/java/com/hazelcast/internal/partition/impl/InternalPartitionServiceImpl.java
@@ -17,18 +17,12 @@
 package com.hazelcast.internal.partition.impl;
 
 import com.hazelcast.cluster.ClusterState;
-import com.hazelcast.core.ExecutionCallback;
-import com.hazelcast.core.HazelcastException;
 import com.hazelcast.core.HazelcastInstanceNotActiveException;
 import com.hazelcast.core.Member;
 import com.hazelcast.core.MemberLeftException;
-import com.hazelcast.core.MigrationEvent;
-import com.hazelcast.core.MigrationEvent.MigrationStatus;
 import com.hazelcast.core.MigrationListener;
 import com.hazelcast.instance.MemberImpl;
 import com.hazelcast.instance.Node;
-import com.hazelcast.instance.NodeState;
-import com.hazelcast.instance.OutOfMemoryErrorDispatcher;
 import com.hazelcast.internal.cluster.MemberInfo;
 import com.hazelcast.internal.cluster.impl.ClusterServiceImpl;
 import com.hazelcast.internal.metrics.Probe;
@@ -37,69 +31,40 @@ import com.hazelcast.internal.partition.InternalPartitionService;
 import com.hazelcast.internal.partition.MigrationInfo;
 import com.hazelcast.internal.partition.PartitionInfo;
 import com.hazelcast.internal.partition.PartitionListener;
-import com.hazelcast.internal.partition.PartitionReplicaChangeReason;
 import com.hazelcast.internal.partition.PartitionRuntimeState;
 import com.hazelcast.internal.partition.PartitionServiceProxy;
-import com.hazelcast.internal.partition.PartitionStateGenerator;
 import com.hazelcast.internal.partition.operation.AssignPartitions;
-import com.hazelcast.internal.partition.operation.ClearReplicaOperation;
-import com.hazelcast.internal.partition.operation.FinalizeMigrationOperation;
+import com.hazelcast.internal.partition.operation.FetchPartitionStateOperation;
 import com.hazelcast.internal.partition.operation.HasOngoingMigration;
-import com.hazelcast.internal.partition.operation.IsReplicaVersionSync;
-import com.hazelcast.internal.partition.operation.MigrationRequestOperation;
 import com.hazelcast.internal.partition.operation.PartitionStateOperation;
-import com.hazelcast.internal.partition.operation.PromoteFromBackupOperation;
-import com.hazelcast.internal.partition.operation.ReplicaSyncRequest;
-import com.hazelcast.internal.partition.operation.ResetReplicaVersionOperation;
-import com.hazelcast.internal.partition.operation.SyncReplicaVersion;
 import com.hazelcast.logging.ILogger;
 import com.hazelcast.logging.Logger;
 import com.hazelcast.nio.Address;
 import com.hazelcast.nio.serialization.Data;
-import com.hazelcast.partition.MigrationEndpoint;
+import com.hazelcast.spi.partition.IPartitionLostEvent;
 import com.hazelcast.partition.NoDataMemberInClusterException;
 import com.hazelcast.partition.PartitionEvent;
 import com.hazelcast.partition.PartitionEventListener;
-import com.hazelcast.partition.PartitionLostEvent;
 import com.hazelcast.partition.PartitionLostListener;
-import com.hazelcast.partition.impl.FetchPartitionStateOperation;
-import com.hazelcast.partition.impl.InternalMigrationListener;
-import com.hazelcast.partition.impl.InternalMigrationListener.MigrationParticipant;
-import com.hazelcast.partition.impl.MigrationCommitOperation;
-import com.hazelcast.partition.membergroup.MemberGroup;
-import com.hazelcast.partition.membergroup.MemberGroupFactory;
-import com.hazelcast.partition.membergroup.MemberGroupFactoryFactory;
 import com.hazelcast.spi.EventPublishingService;
-import com.hazelcast.spi.EventRegistration;
-import com.hazelcast.spi.EventService;
 import com.hazelcast.spi.ExecutionService;
 import com.hazelcast.spi.InvocationBuilder;
 import com.hazelcast.spi.ManagedService;
 import com.hazelcast.spi.NodeEngine;
 import com.hazelcast.spi.Operation;
-import com.hazelcast.spi.OperationResponseHandler;
 import com.hazelcast.spi.OperationService;
 import com.hazelcast.spi.PartitionAwareService;
 import com.hazelcast.spi.TaskScheduler;
 import com.hazelcast.spi.exception.TargetNotMemberException;
 import com.hazelcast.spi.impl.NodeEngineImpl;
-import com.hazelcast.spi.partition.IPartitionLostEvent;
 import com.hazelcast.spi.properties.GroupProperty;
 import com.hazelcast.spi.properties.HazelcastProperties;
-import com.hazelcast.util.Clock;
 import com.hazelcast.util.ExceptionUtil;
 import com.hazelcast.util.FutureUtil.ExceptionHandler;
 import com.hazelcast.util.HashUtil;
-import com.hazelcast.util.Preconditions;
-import com.hazelcast.util.scheduler.CoalescingDelayedTrigger;
-import com.hazelcast.util.scheduler.EntryTaskScheduler;
-import com.hazelcast.util.scheduler.EntryTaskSchedulerFactory;
-import com.hazelcast.util.scheduler.ScheduleType;
 import com.hazelcast.util.scheduler.ScheduledEntry;
-import com.hazelcast.util.scheduler.ScheduledEntryProcessor;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -110,18 +75,13 @@ import java.util.Properties;
 import java.util.Set;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
-import java.util.concurrent.Semaphore;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.AtomicLong;
-import java.util.concurrent.atomic.AtomicReferenceArray;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
 import java.util.logging.Level;
 
 import static com.hazelcast.cluster.memberselector.MemberSelectors.DATA_MEMBER_SELECTOR;
-import static com.hazelcast.spi.impl.OperationResponseHandlerFactory.createErrorLoggingResponseHandler;
 import static com.hazelcast.util.FutureUtil.logAllExceptions;
 import static com.hazelcast.util.FutureUtil.waitWithDeadline;
 import static java.lang.Math.max;
@@ -133,63 +93,31 @@ import static java.lang.Math.min;
 public class InternalPartitionServiceImpl implements InternalPartitionService, ManagedService,
         EventPublishingService<PartitionEvent, PartitionEventListener<PartitionEvent>>, PartitionAwareService {
 
-    private static final String EXCEPTION_MSG_PARTITION_STATE_SYNC_TIMEOUT = "Partition state sync invocation timed out";
-
-    private static final int DEFAULT_PAUSE_MILLIS = 1000;
     private static final int PARTITION_OWNERSHIP_WAIT_MILLIS = 10;
-    private static final int REPLICA_SYNC_CHECK_TIMEOUT_SECONDS = 10;
+    private static final String EXCEPTION_MSG_PARTITION_STATE_SYNC_TIMEOUT = "Partition state sync invocation timed out";
 
     private final Node node;
     private final NodeEngineImpl nodeEngine;
     private final ILogger logger;
+
     private final int partitionCount;
-    private final InternalPartitionImpl[] partitions;
-    private final PartitionReplicaVersions[] replicaVersions;
-    private final AtomicReferenceArray<ReplicaSyncInfo> replicaSyncRequests;
-    private final EntryTaskScheduler<Integer, ReplicaSyncInfo> replicaSyncScheduler;
-    @Probe
-    private final Semaphore replicaSyncProcessLock;
-    private final MigrationThread migrationThread;
-    private final long partitionMigrationInterval;
+
     private final long partitionMigrationTimeout;
-    private final long backupSyncCheckInterval;
-    private final int maxParallelReplications;
-    private final PartitionStateGenerator partitionStateGenerator;
-    private final MemberGroupFactory memberGroupFactory;
+
     private final PartitionServiceProxy proxy;
     private final Lock lock = new ReentrantLock();
     private final InternalPartitionListener partitionListener;
 
-    @Probe
-    private final AtomicInteger stateVersion = new AtomicInteger();
-
-    private final MigrationQueue migrationQueue = new MigrationQueue();
-    private final AtomicBoolean migrationAllowed = new AtomicBoolean(true);
-    @Probe(name = "lastRepartitionTime")
-    private final AtomicLong lastRepartitionTime = new AtomicLong();
-    private final CoalescingDelayedTrigger delayedResumeMigrationTrigger;
+    private final PartitionStateManager partitionStateManager;
+    private final MigrationManager migrationManager;
+    private final PartitionReplicaManager replicaManager;
+    private final PartitionReplicaChecker partitionReplicaChecker;
+    private final PartitionEventManager partitionEventManager;
 
     private final ExceptionHandler partitionStateSyncTimeoutHandler;
 
-    @Probe
-    // can be read and written concurrently...
-    private volatile int memberGroupsSize;
-
-    // updates will be done under lock, but reads will be multithreaded.
-    // set to true when the partitions are assigned for the first time. remains true until partition service has been reset.
-    private volatile boolean initialized;
-
-    // updates will be done under lock, but reads will be multithreaded.
-    private volatile MigrationInfo activeMigrationInfo;
-
-    // both reads and updates will be done under lock!
-    private final LinkedList<MigrationInfo> completedMigrations = new LinkedList<MigrationInfo>();
-
-    @Probe
-    private final AtomicLong completedMigrationCounter = new AtomicLong();
-
-    private volatile InternalMigrationListener internalMigrationListener
-            = new InternalMigrationListener.NopInternalMigrationListener();
+    // used to limit partition assignment requests sent to master
+    private final AtomicBoolean triggerMasterFlag = new AtomicBoolean(false);
 
     private volatile Address lastMaster;
 
@@ -199,124 +127,52 @@ public class InternalPartitionServiceImpl implements InternalPartitionService, M
         this.node = node;
         this.nodeEngine = node.nodeEngine;
         this.logger = node.getLogger(InternalPartitionService.class);
+
+        partitionListener = new InternalPartitionListener(node, this);
+
+        partitionStateManager = new PartitionStateManager(node, this, partitionListener);
+        migrationManager = new MigrationManager(node, this, lock);
+        replicaManager = new PartitionReplicaManager(node, this);
+
+        partitionReplicaChecker = new PartitionReplicaChecker(node, this);
+        partitionEventManager = new PartitionEventManager(node, this);
+
         partitionStateSyncTimeoutHandler =
                 logAllExceptions(logger, EXCEPTION_MSG_PARTITION_STATE_SYNC_TIMEOUT, Level.FINEST);
-        this.partitions = new InternalPartitionImpl[partitionCount];
-        partitionListener = new InternalPartitionListener(this, node.getThisAddress());
-        for (int i = 0; i < partitionCount; i++) {
-            this.partitions[i] = new InternalPartitionImpl(i, partitionListener, node.getThisAddress());
-        }
-        replicaVersions = new PartitionReplicaVersions[partitionCount];
-        for (int i = 0; i < replicaVersions.length; i++) {
-            replicaVersions[i] = new PartitionReplicaVersions(i);
-        }
-
-        memberGroupFactory = MemberGroupFactoryFactory.newMemberGroupFactory(node.getConfig().getPartitionGroupConfig());
-        partitionStateGenerator = new PartitionStateGeneratorImpl();
-
-        long intervalMillis = properties.getMillis(GroupProperty.PARTITION_MIGRATION_INTERVAL);
-        partitionMigrationInterval = (intervalMillis > 0 ? intervalMillis : 0);
 
         partitionMigrationTimeout = properties.getMillis(GroupProperty.PARTITION_MIGRATION_TIMEOUT);
 
-        migrationThread = new MigrationThread(node);
-        proxy = new PartitionServiceProxy(this);
+        proxy = new PartitionServiceProxy(nodeEngine, this);
 
-        ExecutionService executionService = nodeEngine.getExecutionService();
-        TaskScheduler globalScheduler = executionService.getGlobalTaskScheduler();
-
-        // The reason behind this scheduler to have POSTPONE type is as follows:
-        // When a node shifts up in the replica table upon a node failure, it sends a sync request to the partition owner and
-        // registers it to the replicaSyncRequests. If another node fails before the already-running sync process completes,
-        // the new sync request is simply scheduled to a further time. Again, before the already-running sync process completes,
-        // if another node fails for the third time, the already-scheduled sync request should be overwritten with the new one.
-        // This is because this node is shifted up to a higher level when the third node failure occurs and its respective sync
-        // request will inherently include the backup data that is requested by the previously scheduled sync request.
-        replicaSyncScheduler = EntryTaskSchedulerFactory.newScheduler(globalScheduler,
-                new ReplicaSyncEntryProcessor(this), ScheduleType.POSTPONE);
-
-        replicaSyncRequests = new AtomicReferenceArray<ReplicaSyncInfo>(partitionCount);
-
-        long maxMigrationDelayMs = calculateMaxMigrationDelayOnMemberRemoved();
-        long minMigrationDelayMs = calculateMigrationDelayOnMemberRemoved(maxMigrationDelayMs);
-        this.delayedResumeMigrationTrigger = new CoalescingDelayedTrigger(
-                executionService, minMigrationDelayMs, maxMigrationDelayMs, new Runnable() {
-            @Override
-            public void run() {
-                resumeMigration();
-            }
-        });
-
-        long definedBackupSyncCheckInterval = properties.getSeconds(GroupProperty.PARTITION_BACKUP_SYNC_INTERVAL);
-        backupSyncCheckInterval = definedBackupSyncCheckInterval > 0 ? definedBackupSyncCheckInterval : 1;
-        maxParallelReplications = properties.getInteger(GroupProperty.PARTITION_MAX_PARALLEL_REPLICATIONS);
-        replicaSyncProcessLock = new Semaphore(maxParallelReplications);
         nodeEngine.getMetricsRegistry().scanAndRegister(this, "partitions");
-    }
-
-    private long calculateMaxMigrationDelayOnMemberRemoved() {
-        // hard limit for migration pause is half of the call timeout. otherwise we might experience timeouts
-        return node.getProperties().getMillis(GroupProperty.OPERATION_CALL_TIMEOUT_MILLIS) / 2;
-    }
-
-    private long calculateMigrationDelayOnMemberRemoved(long maxDelayMs) {
-        HazelcastProperties properties = node.getProperties();
-        long migrationDelayMs = properties.getMillis(GroupProperty.MIGRATION_MIN_DELAY_ON_MEMBER_REMOVED_SECONDS);
-
-        long connectionErrorDetectionIntervalMs = properties.getMillis(GroupProperty.CONNECTION_MONITOR_INTERVAL)
-                * properties.getInteger(GroupProperty.CONNECTION_MONITOR_MAX_FAULTS) * 5;
-        migrationDelayMs = Math.max(migrationDelayMs, connectionErrorDetectionIntervalMs);
-
-        long heartbeatIntervalMs = properties.getMillis(GroupProperty.HEARTBEAT_INTERVAL_SECONDS);
-        migrationDelayMs = Math.max(migrationDelayMs, heartbeatIntervalMs * 3);
-
-        migrationDelayMs = min(migrationDelayMs, maxDelayMs);
-        return migrationDelayMs;
-    }
-
-    @Probe(name = "migrationActive")
-    private int migrationActiveProbe() {
-        return migrationAllowed.get() ? 1 : 0;
-    }
-
-    @Probe
-    private int localPartitionCount() {
-        int count = 0;
-        for (InternalPartition partition : partitions) {
-            if (partition.isLocal()) {
-                count++;
-            }
-        }
-        return count;
     }
 
     @Override
     public void init(NodeEngine nodeEngine, Properties properties) {
-        migrationThread.start();
-
         int partitionTableSendInterval = node.getProperties().getSeconds(GroupProperty.PARTITION_TABLE_SEND_INTERVAL);
         if (partitionTableSendInterval <= 0) {
             partitionTableSendInterval = 1;
         }
         ExecutionService executionService = nodeEngine.getExecutionService();
-        executionService.scheduleWithRepetition(new SendPartitionRuntimeStateTask(),
+        executionService.scheduleWithRepetition(new PublishPartitionRuntimeStateTask(node, this),
                 partitionTableSendInterval, partitionTableSendInterval, TimeUnit.SECONDS);
 
-        executionService.scheduleWithRepetition(new SyncReplicaVersionTask(),
-                backupSyncCheckInterval, backupSyncCheckInterval, TimeUnit.SECONDS);
+        migrationManager.start();
+        replicaManager.scheduleReplicaVersionSync(executionService);
     }
 
     @Override
     public Address getPartitionOwner(int partitionId) {
-        if (!initialized) {
+        if (!partitionStateManager.isInitialized()) {
             firstArrangement();
         }
-        if (partitions[partitionId].getOwnerOrNull() == null && !node.isMaster() && node.joined()) {
+        final InternalPartition partition = partitionStateManager.getPartitionImpl(partitionId);
+        if (partition.getOwnerOrNull() == null && !node.isMaster()) {
             if (!isClusterFormedByOnlyLiteMembers()) {
-                notifyMasterToAssignPartitions();
+                triggerMasterToAssignPartitions();
             }
         }
-        return partitions[partitionId].getOwnerOrNull();
+        return partition.getOwnerOrNull();
     }
 
     @Override
@@ -345,54 +201,23 @@ public class InternalPartitionServiceImpl implements InternalPartitionService, M
         return owner;
     }
 
-    private boolean isClusterFormedByOnlyLiteMembers() {
-        final ClusterServiceImpl clusterService = node.getClusterService();
-        return clusterService.getMembers(DATA_MEMBER_SELECTOR).isEmpty();
-    }
-
-    private void notifyMasterToAssignPartitions() {
-        if (initialized) {
-            return;
-        }
-
-        ClusterState clusterState = node.getClusterService().getClusterState();
-        if (clusterState != ClusterState.ACTIVE) {
-            logger.warning("Partitions can't be assigned since cluster-state= " + clusterState);
-            return;
-        }
-
-        if (lock.tryLock()) {
-            try {
-                if (!initialized && !node.isMaster() && node.getMasterAddress() != null && node.joined()) {
-                    Future f = nodeEngine.getOperationService().createInvocationBuilder(SERVICE_NAME, new AssignPartitions(),
-                            node.getMasterAddress()).setTryCount(1).invoke();
-                    f.get(1, TimeUnit.SECONDS);
-                }
-            } catch (Exception e) {
-                logger.finest(e);
-            } finally {
-                lock.unlock();
-            }
-        }
-    }
-
     @Override
     public void firstArrangement() {
-        if (initialized) {
+        if (partitionStateManager.isInitialized()) {
             return;
         }
 
         if (!node.isMaster()) {
-            notifyMasterToAssignPartitions();
+            triggerMasterToAssignPartitions();
             return;
         }
 
         lock.lock();
         try {
-            if (initialized) {
+            if (partitionStateManager.isInitialized()) {
                 return;
             }
-            if (!initializePartitionAssignments()) {
+            if (!partitionStateManager.initializePartitionAssignments()) {
                 return;
             }
             publishPartitionRuntimeState();
@@ -401,98 +226,56 @@ public class InternalPartitionServiceImpl implements InternalPartitionService, M
         }
     }
 
-    private Collection<MemberGroup> createMemberGroups() {
-        final Collection<Member> members = node.getClusterService().getMembers(DATA_MEMBER_SELECTOR);
-        return memberGroupFactory.createMemberGroups(members);
-    }
+    private void triggerMasterToAssignPartitions() {
+        if (partitionStateManager.isInitialized()) {
+            return;
+        }
 
-    private boolean initializePartitionAssignments() {
+        if (!node.joined()) {
+            return;
+        }
+
         ClusterState clusterState = node.getClusterService().getClusterState();
         if (clusterState != ClusterState.ACTIVE) {
             logger.warning("Partitions can't be assigned since cluster-state= " + clusterState);
-            return false;
+            return;
         }
 
-        PartitionStateGenerator psg = partitionStateGenerator;
-        Collection<MemberGroup> memberGroups = createMemberGroups();
-        if (memberGroups.isEmpty()) {
-            logger.warning("No member group is available to assign partition ownership...");
-            return false;
+        if (!triggerMasterFlag.compareAndSet(false, true)) {
+            return;
         }
 
-        logger.info("Initializing cluster partition table arrangement...");
-        Address[][] newState = psg.initialize(memberGroups, partitionCount);
-        if (newState.length != partitionCount) {
-            throw new HazelcastException("Invalid partition count! "
-                    + "Expected: " + partitionCount + ", Actual: " + newState.length);
+        try {
+            final Address masterAddress = node.getMasterAddress();
+            if (masterAddress != null && !masterAddress.equals(node.getThisAddress())) {
+                Future f = nodeEngine.getOperationService().createInvocationBuilder(SERVICE_NAME, new AssignPartitions(),
+                        masterAddress).setTryCount(1).invoke();
+                f.get(1, TimeUnit.SECONDS);
+            }
+        } catch (Exception e) {
+            logger.finest(e);
+        } finally {
+            triggerMasterFlag.set(false);
         }
+    }
 
-        // increment state version to make fail cluster state transaction
-        // if it's started and not locked the state yet.
-        stateVersion.incrementAndGet();
-        clusterState = node.getClusterService().getClusterState();
-        if (clusterState != ClusterState.ACTIVE) {
-            // cluster state is either changed or locked, decrement version back and fail.
-            stateVersion.decrementAndGet();
-            logger.warning("Partitions can't be assigned since cluster-state= " + clusterState);
-            return false;
-        }
-
-        for (int partitionId = 0; partitionId < partitionCount; partitionId++) {
-            InternalPartitionImpl partition = partitions[partitionId];
-            Address[] replicas = newState[partitionId];
-            partition.setReplicaAddresses(replicas);
-        }
-        initialized = true;
-        return true;
+    boolean isClusterFormedByOnlyLiteMembers() {
+        final ClusterServiceImpl clusterService = node.getClusterService();
+        return clusterService.getMembers(DATA_MEMBER_SELECTOR).isEmpty();
     }
 
     public void setInitialState(Address[][] newState, int partitionStateVersion) {
         lock.lock();
         try {
-            if (initialized) {
-                throw new IllegalStateException("Partition table is already initialized!");
-            }
-            logger.info("Setting cluster partition table ...");
-            boolean foundReplica = false;
-            for (int partitionId = 0; partitionId < partitionCount; partitionId++) {
-                InternalPartitionImpl partition = partitions[partitionId];
-                Address[] replicas = newState[partitionId];
-                if (!foundReplica && replicas != null) {
-                    for (int i = 0; i < InternalPartition.MAX_REPLICA_COUNT; i++) {
-                        foundReplica |= replicas[i] != null;
-                    }
-                }
-                partition.setInitialReplicaAddresses(replicas);
-            }
-            stateVersion.set(partitionStateVersion);
-            initialized = foundReplica;
+            partitionStateManager.setInitialState(newState, partitionStateVersion);
         } finally {
             lock.unlock();
         }
     }
 
-    private void updateMemberGroupsSize() {
-        final Collection<MemberGroup> groups = createMemberGroups();
-        int size = 0;
-        for (MemberGroup group : groups) {
-            if (group.size() > 0) {
-                size++;
-            }
-        }
-        memberGroupsSize = size;
-    }
-
     @Override
     public int getMemberGroupsSize() {
-        int size = memberGroupsSize;
-        if (size > 0) {
-            return size;
-        }
-
-        // size = 0 means service is not initialized yet.
-        // return 1 if current node is a data member since there should be at least one member group
-        return node.isLiteMember() ? 0 : 1;
+        return partitionStateManager.getMemberGroupsSize();
     }
 
     @Probe(name = "maxBackupCount")
@@ -504,19 +287,18 @@ public class InternalPartitionServiceImpl implements InternalPartitionService, M
     @Override
     public void memberAdded(MemberImpl member) {
         if (!member.localMember()) {
-            updateMemberGroupsSize();
+            partitionStateManager.updateMemberGroupsSize();
         }
         lastMaster = node.getMasterAddress();
 
         if (node.isMaster()) {
             lock.lock();
             try {
-                migrationQueue.clear();
-                if (initialized) {
+                if (partitionStateManager.isInitialized()) {
                     final ClusterState clusterState = nodeEngine.getClusterService().getClusterState();
                     if (clusterState == ClusterState.ACTIVE) {
-                        stateVersion.incrementAndGet();
-                        migrationQueue.add(new RepartitioningTask());
+                        partitionStateManager.incrementVersion();
+                        migrationManager.triggerRepartitioning();
                     }
 
                     // send initial partition table to newly joined node.
@@ -532,126 +314,52 @@ public class InternalPartitionServiceImpl implements InternalPartitionService, M
     @Override
     public void memberRemoved(final MemberImpl member) {
         logger.info("Removing " + member);
-        updateMemberGroupsSize();
+        partitionStateManager.updateMemberGroupsSize();
 
         final Address deadAddress = member.getAddress();
         final Address thisAddress = node.getThisAddress();
 
-        // TODO: not needed!
-//        if (thisAddress.equals(deadAddress)) {
-//            return;
-//        }
-
         lock.lock();
         try {
             // TODO: why not increment only on master and publish it?
-            if (initialized && node.getClusterService().getClusterState() == ClusterState.ACTIVE) {
-                stateVersion.incrementAndGet();
+            if (partitionStateManager.isInitialized() && node.getClusterService().getClusterState() == ClusterState.ACTIVE) {
+                partitionStateManager.incrementVersion();
             }
 
-            // TODO: should not clear all! but only migration tasks? only to/from dead member?
-            migrationQueue.clear();
+            migrationManager.onMemberRemove(member);
 
             if (node.isMaster() && !thisAddress.equals(lastMaster)) {
-                Runnable runnable = new FixPartitionTableTask();
-                migrationQueue.add(runnable);
+                Runnable runnable = new FetchMostRecentPartitionTableTask();
+                migrationManager.execute(runnable);
             }
-
             lastMaster = node.getMasterAddress();
 
-//            // TODO: why not just rollback them? unless it's committed by target
-//            invalidateActiveMigrationsBelongingTo(deadAddress);
-
-            // TODO: still need to pause?
+            // TODO: here or before all other actions?
             // Pause migration and let all other members notice the dead member
-            // and fix their own partitions.
             // Otherwise new master may take action fast and send new partition state
             // before other members realize the dead one.
-//            pauseMigration();
+            migrationManager.pauseMigration();
 
             // TODO: this looks fine, a local optimization
-            cancelReplicaSyncRequestsInternal(deadAddress);
+            replicaManager.cancelReplicaSyncRequestsTo(deadAddress);
 
             if (node.isMaster()) {
-                migrationQueue.add(new Runnable() {
-                    @Override
-                    public void run() {
-                        if (!initialized) {
-                            return;
-                        }
-                        // invalidate, migration will be eventually rolled back
-                        invalidateActiveMigrationsBelongingTo(deadAddress);
-
-                        removeDeadAddress(deadAddress);
-
-                        syncPartitionRuntimeState();
-                    }
-                });
-
-                migrationQueue.add(new RepartitioningTask());
+                migrationManager.execute(new RepairPartitionTableTask(deadAddress));
+                migrationManager.triggerRepartitioning();
             }
 
-//            resumeMigrationEventually();
+            migrationManager.resumeMigrationEventually();
         } finally {
             lock.unlock();
-        }
-    }
-
-    private void invalidateActiveMigrationsBelongingTo(Address deadAddress) {
-        if (activeMigrationInfo != null) {
-            if (deadAddress.equals(activeMigrationInfo.getSource())
-                    || deadAddress.equals(activeMigrationInfo.getDestination())) {
-                activeMigrationInfo.invalidate();
-            }
         }
     }
 
     public void cancelReplicaSyncRequestsTo(Address deadAddress) {
         lock.lock();
         try {
-            cancelReplicaSyncRequestsInternal(deadAddress);
+            replicaManager.cancelReplicaSyncRequestsTo(deadAddress);
         } finally {
             lock.unlock();
-        }
-    }
-
-    private void cancelReplicaSyncRequestsInternal(Address deadAddress) {
-        for (int partitionId = 0; partitionId < partitionCount; partitionId++) {
-            ReplicaSyncInfo syncInfo = replicaSyncRequests.get(partitionId);
-            if (syncInfo != null && deadAddress.equals(syncInfo.target)) {
-                cancelReplicaSync(partitionId);
-            }
-        }
-    }
-
-    public void cancelReplicaSync(int partitionId) {
-        ReplicaSyncInfo syncInfo = replicaSyncRequests.get(partitionId);
-        if (syncInfo != null && replicaSyncRequests.compareAndSet(partitionId, syncInfo, null)) {
-            replicaSyncScheduler.cancel(partitionId);
-            releaseReplicaSyncPermit();
-        }
-    }
-
-    private void resumeMigrationEventually() {
-        delayedResumeMigrationTrigger.executeWithDelay();
-    }
-
-    private void removeDeadAddress(Address deadAddress) {
-        for (InternalPartitionImpl partition : partitions) {
-//            if (deadAddress.equals(partition.getOwnerOrNull()) && thisAddress.equals(partition.getReplicaAddress(1))) {
-//                partition.setMigrating(true);
-//            }
-
-            // shift partition table up.
-//            partition.onDeadAddress(deadAddress);
-
-            // safety check!
-//            if (partition.onDeadAddress(deadAddress)) {
-//                throw new IllegalStateException("Duplicate address found in partition replicas!");
-//            }
-
-            // set null for replica assignments to this address
-            partition.removeAddress(deadAddress);
         }
     }
 
@@ -660,7 +368,7 @@ public class InternalPartitionServiceImpl implements InternalPartitionService, M
     }
 
     private PartitionRuntimeState createPartitionState(Collection<MemberImpl> members) {
-        if (!initialized) {
+        if (!partitionStateManager.isInitialized()) {
             return null;
         }
 
@@ -671,19 +379,22 @@ public class InternalPartitionServiceImpl implements InternalPartitionService, M
                 MemberInfo memberInfo = new MemberInfo(member.getAddress(), member.getUuid(), member.getAttributes());
                 memberInfos.add(memberInfo);
             }
-            ArrayList<MigrationInfo> migrationInfos = new ArrayList<MigrationInfo>(completedMigrations);
+            List<MigrationInfo> completedMigrations = migrationManager.getCompletedMigrations();
             ILogger logger = node.getLogger(PartitionRuntimeState.class);
+
+            InternalPartition[] partitions = partitionStateManager.getPartitions();
+
             PartitionRuntimeState state =
-                    new PartitionRuntimeState(logger, memberInfos, partitions, migrationInfos, stateVersion.get());
-            state.setActiveMigration(activeMigrationInfo);
+                    new PartitionRuntimeState(logger, memberInfos, partitions, completedMigrations, getPartitionStateVersion());
+            state.setActiveMigration(migrationManager.getActiveMigration());
             return state;
         } finally {
             lock.unlock();
         }
     }
 
-    private PartitionRuntimeState createMigrationCommitPartitionState(MigrationInfo migrationInfo, Address[] newAddresses) {
-        if (!initialized) {
+    PartitionRuntimeState createMigrationCommitPartitionState(MigrationInfo migrationInfo, Address[] newAddresses) {
+        if (!partitionStateManager.isInitialized()) {
             return null;
         }
         Collection<MemberImpl> members = node.clusterService.getMemberImpls();
@@ -694,25 +405,22 @@ public class InternalPartitionServiceImpl implements InternalPartitionService, M
                 MemberInfo memberInfo = new MemberInfo(member.getAddress(), member.getUuid(), member.getAttributes());
                 memberInfos.add(memberInfo);
             }
-            ArrayList<MigrationInfo> completedMigrations = new ArrayList<MigrationInfo>(this.completedMigrations);
+            List<MigrationInfo> completedMigrations = migrationManager.getCompletedMigrations();
             ILogger logger = node.getLogger(PartitionRuntimeState.class);
 
-            InternalPartitionImpl[] partitions = new InternalPartitionImpl[partitionCount];
-            for (int i = 0; i < partitionCount; i++) {
-                partitions[i] = this.partitions[i].copy();
-            }
+            InternalPartition[] partitions = partitionStateManager.getPartitionsCopy();
 
             int partitionId = migrationInfo.getPartitionId();
-            partitions[partitionId].setReplicaAddresses(newAddresses);
+            partitionStateManager.setReplicaAddresses(partitions[partitionId], newAddresses);
 
-            return new PartitionRuntimeState(logger, memberInfos, partitions, completedMigrations, stateVersion.get() + 1);
+            return new PartitionRuntimeState(logger, memberInfos, partitions, completedMigrations, getPartitionStateVersion() + 1);
         } finally {
             lock.unlock();
         }
     }
 
-    private void publishPartitionRuntimeState() {
-        if (!initialized) {
+    void publishPartitionRuntimeState() {
+        if (!partitionStateManager.isInitialized()) {
             // do not send partition state until initialized!
             return;
         }
@@ -748,42 +456,12 @@ public class InternalPartitionServiceImpl implements InternalPartitionService, M
         }
     }
 
-    private void syncPartitionRuntimeState() {
+    void syncPartitionRuntimeState() {
         syncPartitionRuntimeState(node.clusterService.getMemberImpls());
     }
 
-    private boolean commitMigrationToDestination(MigrationInfo migrationInfo, Address[] newAddresses) {
-        if (!node.isMaster()) {
-            return false;
-        }
-
-        if (node.getThisAddress().equals(migrationInfo.getDestination())) {
-            return true;
-        }
-
-        try {
-            PartitionRuntimeState partitionState = createMigrationCommitPartitionState(migrationInfo, newAddresses);
-            MigrationCommitOperation operation = new MigrationCommitOperation(partitionState);
-            Future<Boolean> future = nodeEngine.getOperationService()
-                                                             .createInvocationBuilder(SERVICE_NAME, operation,
-                                                                     migrationInfo.getDestination())
-                                                             .setTryCount(Integer.MAX_VALUE)
-                                                             .setCallTimeout(Long.MAX_VALUE).invoke();
-            future.get();
-            return true;
-        } catch (Throwable t) {
-            if (t instanceof MemberLeftException || t instanceof TargetNotMemberException) {
-                logger.warning("Migration commit failed for " + migrationInfo + " since destination left the cluster");
-            } else {
-                logger.severe("Migration commit failed for " + migrationInfo + " with " + t);
-            }
-
-            return false;
-        }
-    }
-
-    private void syncPartitionRuntimeState(Collection<MemberImpl> members) {
-        if (!initialized) {
+    void syncPartitionRuntimeState(Collection<MemberImpl> members) {
+        if (!partitionStateManager.isInitialized()) {
             // do not send partition state until initialized!
             return;
         }
@@ -860,8 +538,8 @@ public class InternalPartitionServiceImpl implements InternalPartitionService, M
     }
 
     private void applyNewState(PartitionRuntimeState partitionState, Address sender) {
-        stateVersion.set(partitionState.getVersion());
-        initialized = true;
+        partitionStateManager.setVersion(partitionState.getVersion());
+        partitionStateManager.setInitialized();
 
         PartitionInfo[] state = partitionState.getPartitions();
         filterAndLogUnknownAddressesInPartitionTable(sender, state);
@@ -871,7 +549,7 @@ public class InternalPartitionServiceImpl implements InternalPartitionService, M
     private void finalizeOrRollbackMigration(PartitionRuntimeState partitionState, PartitionInfo[] state) {
         Collection<MigrationInfo> completedMigrations = partitionState.getCompletedMigrations();
         for (MigrationInfo completedMigration : completedMigrations) {
-            addCompletedMigration(completedMigration);
+            migrationManager.addCompletedMigration(completedMigration);
             int partitionId = completedMigration.getPartitionId();
             PartitionInfo partitionInfo = state[partitionId];
             // mdogan:
@@ -882,22 +560,17 @@ public class InternalPartitionServiceImpl implements InternalPartitionService, M
             // For the time being, leaving this stuff as is to not to change behaviour.
 
             // TODO BASRI IS THIS STILL NECESSARY? CAN WE MOVE UPDATEALLPARTITIONS(state) BEFORE FOR LOOP AND REMOVE NEXT LINE?
-            updatePartitionReplicaAddresses(partitionInfo.getPartitionId(), partitionInfo.getReplicaAddresses());
-            scheduleActiveMigrationFinalization(completedMigration);
+            partitionStateManager.updateReplicaAddresses(partitionInfo.getPartitionId(), partitionInfo.getReplicaAddresses());
+            migrationManager.scheduleActiveMigrationFinalization(completedMigration);
         }
 
         updateAllPartitions(state);
     }
 
-    private void updatePartitionReplicaAddresses(int partitionId, Address[] replicaAddresses) {
-        InternalPartitionImpl partition = partitions[partitionId];
-        partition.setReplicaAddresses(replicaAddresses);
-    }
-
     private void updateAllPartitions(PartitionInfo[] state) {
         for (int partitionId = 0; partitionId < partitionCount; partitionId++) {
             final PartitionInfo partitionInfo = state[partitionId];
-            updatePartitionReplicaAddresses(partitionInfo.getPartitionId(), partitionInfo.getReplicaAddresses());
+            partitionStateManager.updateReplicaAddresses(partitionInfo.getPartitionId(), partitionInfo.getReplicaAddresses());
         }
     }
 
@@ -944,274 +617,16 @@ public class InternalPartitionServiceImpl implements InternalPartitionService, M
         }
     }
 
-    private void scheduleActiveMigrationFinalization(final MigrationInfo migrationInfo) {
-        lock.lock();
-        try {
-            if (migrationInfo.equals(activeMigrationInfo)) {
-                if (activeMigrationInfo.startProcessing()) {
-                    finalizeMigrationInfo(activeMigrationInfo);
-                } else {
-                    logger.info("Scheduling finalization of " + migrationInfo
-                            + ", because migration process is currently running.");
-                    nodeEngine.getExecutionService().schedule(new Runnable() {
-                        @Override
-                        public void run() {
-                            scheduleActiveMigrationFinalization(migrationInfo);
-                        }
-                    }, 3, TimeUnit.SECONDS);
-                }
-            }
-        } finally {
-            lock.unlock();
-        }
-    }
-
-    private void finalizeMigrationInfo(MigrationInfo migrationInfo) {
-        try {
-            Address thisAddress = node.getThisAddress();
-            boolean source = thisAddress.equals(migrationInfo.getSource());
-            boolean destination = thisAddress.equals(migrationInfo.getDestination());
-            if (source || destination) {
-                int partitionId = migrationInfo.getPartitionId();
-                InternalPartitionImpl migratingPartition = getPartitionImpl(partitionId);
-                Address ownerAddress = migratingPartition.getOwnerOrNull();
-                boolean success = migrationInfo.getDestination().equals(ownerAddress);
-                MigrationParticipant participant = source ? MigrationParticipant.SOURCE : MigrationParticipant.DESTINATION;
-                if (success) {
-                    internalMigrationListener.onMigrationCommit(participant, migrationInfo);
-                } else {
-                    internalMigrationListener.onMigrationRollback(participant, migrationInfo);
-                }
-
-                MigrationEndpoint endpoint = source ? MigrationEndpoint.SOURCE : MigrationEndpoint.DESTINATION;
-                FinalizeMigrationOperation op = new FinalizeMigrationOperation(endpoint, success);
-                op.setPartitionId(partitionId)
-                        .setNodeEngine(nodeEngine)
-                        .setValidateTarget(false)
-                        .setService(this);
-                nodeEngine.getOperationService().executeOperation(op);
-            } else {
-                logger.severe("Failed to finalize migration because this member " + thisAddress
-                        + " is not a participant of the migration: " + migrationInfo);
-            }
-        } catch (Exception e) {
-            logger.warning(e);
-        } finally {
-            migrationInfo.doneProcessing();
-        }
-    }
-
-    public boolean addActiveMigration(MigrationInfo migrationInfo) {
-        lock.lock();
-        try {
-            if (activeMigrationInfo == null) {
-                partitions[migrationInfo.getPartitionId()].setMigrating(true);
-                activeMigrationInfo = migrationInfo;
-                return true;
-            }
-
-            logger.warning(migrationInfo + " not added! Already existing active migration: " + activeMigrationInfo);
-            return false;
-        } finally {
-            lock.unlock();
-        }
-    }
-
-    private MemberImpl getMasterMember() {
-        return node.clusterService.getMember(node.getMasterAddress());
-    }
-
-    public MigrationInfo getActiveMigration(int partitionId) {
-        MigrationInfo activeMigrationInfo = this.activeMigrationInfo;
-        if (activeMigrationInfo != null && activeMigrationInfo.getPartitionId() == partitionId) {
-            return activeMigrationInfo;
-        }
-
-        return null;
-    }
-
-    public boolean removeActiveMigration(int partitionId) {
-        boolean success = false;
-        lock.lock();
-        try {
-            if (activeMigrationInfo != null) {
-                if (activeMigrationInfo.getPartitionId() == partitionId) {
-                    partitions[partitionId].setMigrating(false);
-                    this.activeMigrationInfo = null;
-                    success = true;
-                } else if (logger.isFinestEnabled()) {
-                    logger.finest("active migration not remove because it has different partitionId! partitionId=" + partitionId
-                    + " active migration=" + activeMigrationInfo);
-                }
-            }
-        } finally {
-            lock.unlock();
-        }
-
-        return success;
-    }
-
-    private void addCompletedMigration(MigrationInfo migrationInfo) {
-        completedMigrationCounter.incrementAndGet();
-
-        lock.lock();
-        try {
-            if (completedMigrations.size() > 25) {
-                completedMigrations.removeFirst();
-            }
-            completedMigrations.add(migrationInfo);
-        } finally {
-            lock.unlock();
-        }
-    }
-
-    private void evictCompletedMigrations() {
-        lock.lock();
-        try {
-            if (!completedMigrations.isEmpty()) {
-                completedMigrations.removeFirst();
-            }
-        } finally {
-            lock.unlock();
-        }
-    }
-
-    // This method is called in backup node. Given all other conditions are satisfied,
-    // this method initiates a replica sync operation and registers it to replicaSyncRequest.
-    // If another sync request is already registered, it schedules the new replica sync request to a further time.
-    public void triggerPartitionReplicaSync(int partitionId, int replicaIndex, long delayMillis) {
-        if (replicaIndex < 0 || replicaIndex > InternalPartition.MAX_REPLICA_COUNT) {
-            throw new IllegalArgumentException("Invalid replica index! replicaIndex=" + replicaIndex
-                    + " for partitionId=" + partitionId);
-        }
-
-        if (!checkSyncPartitionTarget(partitionId, replicaIndex)) {
-            return;
-        }
-
-        InternalPartitionImpl partition = getPartitionImpl(partitionId);
-        Address target = partition.getOwnerOrNull();
-        ReplicaSyncInfo syncInfo = new ReplicaSyncInfo(partitionId, replicaIndex, target);
-
-        if (delayMillis > 0) {
-            schedulePartitionReplicaSync(syncInfo, target, delayMillis, "EXPLICIT DELAY");
-            return;
-        }
-
-        // mdogan:
-        // merged two conditions into single `if-return` block to
-        // conform checkstyle return-count rule.
-        if (!isReplicaSyncAllowed() || partition.isMigrating()) {
-
-            schedulePartitionReplicaSync(syncInfo, target, REPLICA_SYNC_RETRY_DELAY,
-                    "MIGRATION IS DISABLED OR PARTITION IS MIGRATING");
-            return;
-        }
-
-        if (replicaSyncRequests.compareAndSet(partitionId, null, syncInfo)) {
-            if (fireSyncReplicaRequest(syncInfo, target)) {
-                return;
-            }
-
-            replicaSyncRequests.compareAndSet(partitionId, syncInfo, null);
-            schedulePartitionReplicaSync(syncInfo, target, REPLICA_SYNC_RETRY_DELAY, "NO PERMIT AVAILABLE");
-            return;
-        }
-
-        long scheduleDelay = getReplicaSyncScheduleDelay(partitionId);
-        schedulePartitionReplicaSync(syncInfo, target, scheduleDelay, "ANOTHER SYNC IN PROGRESS");
-    }
-
-    private long getReplicaSyncScheduleDelay(int partitionId) {
-        long scheduleDelay = DEFAULT_REPLICA_SYNC_DELAY;
-        Address thisAddress = node.getThisAddress();
-        InternalPartitionImpl partition = getPartitionImpl(partitionId);
-        ReplicaSyncInfo currentSyncInfo = replicaSyncRequests.get(partitionId);
-        if (currentSyncInfo != null
-                && !thisAddress.equals(partition.getReplicaAddress(currentSyncInfo.replicaIndex))) {
-            clearReplicaSyncRequest(partitionId, currentSyncInfo.replicaIndex);
-            scheduleDelay = REPLICA_SYNC_RETRY_DELAY;
-        }
-        return scheduleDelay;
-    }
-
-    private boolean fireSyncReplicaRequest(ReplicaSyncInfo syncInfo, Address target) {
-        if (node.clusterService.isMemberRemovedWhileClusterIsNotActive(target)) {
-            return false;
-        }
-
-        if (tryToAcquireReplicaSyncPermit()) {
-            int partitionId = syncInfo.partitionId;
-            int replicaIndex = syncInfo.replicaIndex;
-            replicaSyncScheduler.cancel(partitionId);
-
-            if (logger.isFinestEnabled()) {
-                logger.finest("Sending sync replica request to -> " + target + "; for partitionId=" + partitionId
-                        + ", replicaIndex=" + replicaIndex);
-            }
-            replicaSyncScheduler.schedule(partitionMigrationTimeout, partitionId, syncInfo);
-            ReplicaSyncRequest syncRequest = new ReplicaSyncRequest(partitionId, replicaIndex);
-            nodeEngine.getOperationService().send(syncRequest, target);
-            return true;
-        }
-        return false;
-    }
-
-    private void schedulePartitionReplicaSync(ReplicaSyncInfo syncInfo, Address target, long delayMillis, String reason) {
-        int partitionId = syncInfo.partitionId;
-        int replicaIndex = syncInfo.replicaIndex;
-
-        if (logger.isFinestEnabled()) {
-            logger.finest(
-                    "Scheduling [" + delayMillis + "ms] sync replica request to -> " + target + "; for partitionId=" + partitionId
-                            + ", replicaIndex=" + replicaIndex + ". Reason: [" + reason + "]");
-        }
-        replicaSyncScheduler.schedule(delayMillis, partitionId, syncInfo);
-    }
-
-    private boolean checkSyncPartitionTarget(int partitionId, int replicaIndex) {
-        final InternalPartitionImpl partition = getPartitionImpl(partitionId);
-        final Address target = partition.getOwnerOrNull();
-        if (target == null) {
-            logger.info("Sync replica target is null, no need to sync -> partitionId=" + partitionId + ", replicaIndex="
-                    + replicaIndex);
-            return false;
-        }
-
-        Address thisAddress = nodeEngine.getThisAddress();
-        if (target.equals(thisAddress)) {
-            if (logger.isFinestEnabled()) {
-                logger.finest("This node is now owner of partition, cannot sync replica -> partitionId=" + partitionId
-                        + ", replicaIndex=" + replicaIndex + ", partition-info="
-                        + getPartitionImpl(partitionId));
-            }
-            return false;
-        }
-
-        if (!partition.isOwnerOrBackup(thisAddress)) {
-            if (logger.isFinestEnabled()) {
-                logger.finest("This node is not backup replica of partitionId=" + partitionId
-                        + ", replicaIndex=" + replicaIndex + " anymore.");
-            }
-            return false;
-        }
-        return true;
-    }
-
     @Override
     public InternalPartition[] getPartitions() {
         //a defensive copy is made to prevent breaking with the old approach, but imho not needed
-        InternalPartition[] result = new InternalPartition[partitions.length];
-        System.arraycopy(partitions, 0, result, 0, partitions.length);
+        InternalPartition[] result = new InternalPartition[partitionCount];
+        System.arraycopy(partitionStateManager.getPartitions(), 0, result, 0, partitionCount);
         return result;
     }
 
-    public MemberImpl getMember(Address address) {
+    MemberImpl getMember(Address address) {
         return node.clusterService.getMember(address);
-    }
-
-    public InternalPartitionImpl getPartitionImpl(int partitionId) {
-        return partitions[partitionId];
     }
 
     @Override
@@ -1221,7 +636,7 @@ public class InternalPartitionServiceImpl implements InternalPartitionService, M
 
     @Override
     public InternalPartition getPartition(int partitionId, boolean triggerOwnerAssignment) {
-        InternalPartitionImpl p = partitions[partitionId];
+        InternalPartitionImpl p = partitionStateManager.getPartitionImpl(partitionId);
         if (triggerOwnerAssignment && p.getOwnerOrNull() == null) {
             // probably ownerships are not set yet.
             // force it.
@@ -1232,88 +647,19 @@ public class InternalPartitionServiceImpl implements InternalPartitionService, M
 
     @Override
     public boolean prepareToSafeShutdown(long timeout, TimeUnit unit) {
-        long timeoutInMillis = unit.toMillis(timeout);
-        long sleep = DEFAULT_PAUSE_MILLIS;
-        while (timeoutInMillis > 0) {
-            while (timeoutInMillis > 0 && shouldWaitMigrationOrBackups(Level.INFO)) {
-                timeoutInMillis = sleepWithBusyWait(timeoutInMillis, sleep);
-            }
-            if (timeoutInMillis <= 0) {
-                break;
-            }
-
-            if (node.isMaster()) {
-                final List<MemberImpl> members = getCurrentMembersAndMembersRemovedWhileNotClusterNotActive();
-                syncPartitionRuntimeState(members);
-            } else {
-                timeoutInMillis = waitForOngoingMigrations(timeoutInMillis, sleep);
-                if (timeoutInMillis <= 0) {
-                    break;
-                }
-            }
-
-            long start = Clock.currentTimeMillis();
-            boolean ok = checkReplicaSyncState();
-            timeoutInMillis -= (Clock.currentTimeMillis() - start);
-            if (ok) {
-                logger.finest("Replica sync state before shutdown is OK");
-                return true;
-            } else {
-                if (timeoutInMillis <= 0) {
-                    break;
-                }
-                logger.info("Some backup replicas are inconsistent with primary, waiting for synchronization. Timeout: "
-                        + timeoutInMillis + "ms");
-                timeoutInMillis = sleepWithBusyWait(timeoutInMillis, sleep);
-            }
-        }
-        return false;
+        return partitionReplicaChecker.prepareToSafeShutdown(timeout, unit);
     }
 
-    private List<MemberImpl> getCurrentMembersAndMembersRemovedWhileNotClusterNotActive() {
+    List<MemberImpl> getCurrentMembersAndMembersRemovedWhileNotClusterNotActive() {
         final List<MemberImpl> members = new ArrayList<MemberImpl>();
         members.addAll(node.clusterService.getMemberImpls());
         members.addAll(node.clusterService.getMembersRemovedWhileClusterIsNotActive());
         return members;
     }
 
-    private long waitForOngoingMigrations(long timeoutInMillis, long sleep) {
-        long timeout = timeoutInMillis;
-        while (timeout > 0 && hasOnGoingMigrationMaster(Level.WARNING)) {
-            // ignore elapsed time during master inv.
-            logger.info("Waiting for the master node to complete remaining migrations!");
-            timeout = sleepWithBusyWait(timeout, sleep);
-        }
-        return timeout;
-    }
-
-    private long sleepWithBusyWait(long timeoutInMillis, long sleep) {
-        try {
-            //noinspection BusyWait
-            Thread.sleep(sleep);
-        } catch (InterruptedException ie) {
-            logger.finest("Busy wait interrupted", ie);
-        }
-        return timeoutInMillis - sleep;
-    }
-
     @Override
     public boolean isMemberStateSafe() {
-        return getMemberState() == InternalPartitionServiceState.SAFE;
-    }
-
-    public InternalPartitionServiceState getMemberState() {
-        if (hasOnGoingMigrationLocal()) {
-            return InternalPartitionServiceState.MIGRATION_LOCAL;
-        }
-
-        if (!node.isMaster()) {
-            if (hasOnGoingMigrationMaster(Level.OFF)) {
-                return InternalPartitionServiceState.MIGRATION_ON_MASTER;
-            }
-        }
-
-        return isReplicaInSyncState() ? InternalPartitionServiceState.SAFE : InternalPartitionServiceState.REPLICA_NOT_SYNC;
+        return partitionReplicaChecker.getMemberState() == InternalPartitionServiceState.SAFE;
     }
 
     @Override
@@ -1343,224 +689,7 @@ public class InternalPartitionServiceImpl implements InternalPartitionService, M
 
     @Override
     public boolean hasOnGoingMigrationLocal() {
-        return activeMigrationInfo != null || migrationQueue.isNonEmpty()
-                || migrationQueue.hasMigrationTasks();
-    }
-
-    private boolean isReplicaInSyncState() {
-        if (!initialized || !hasMultipleMemberGroups()) {
-            return true;
-        }
-        final int replicaIndex = 1;
-        final List<Future> futures = new ArrayList<Future>();
-        final Address thisAddress = node.getThisAddress();
-        for (InternalPartitionImpl partition : partitions) {
-            final Address owner = partition.getOwnerOrNull();
-            if (thisAddress.equals(owner)) {
-                if (partition.getReplicaAddress(replicaIndex) != null) {
-                    final int partitionId = partition.getPartitionId();
-                    final long replicaVersion = getCurrentReplicaVersion(replicaIndex, partitionId);
-                    final Operation operation = createReplicaSyncStateOperation(replicaVersion, partitionId);
-                    final Future future = invoke(operation, replicaIndex, partitionId);
-                    futures.add(future);
-                }
-            }
-        }
-        if (futures.isEmpty()) {
-            return true;
-        }
-        for (Future future : futures) {
-            boolean isSync = getFutureResultSilently(future, REPLICA_SYNC_CHECK_TIMEOUT_SECONDS, TimeUnit.SECONDS);
-            if (!isSync) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    private long getCurrentReplicaVersion(int replicaIndex, int partitionId) {
-        final long[] versions = getPartitionReplicaVersions(partitionId);
-        return versions[replicaIndex - 1];
-    }
-
-    private boolean getFutureResultSilently(Future future, long seconds, TimeUnit unit) {
-        boolean sync;
-        try {
-            sync = (Boolean) future.get(seconds, unit);
-        } catch (Throwable t) {
-            sync = false;
-            logger.finest("Exception while getting future", t);
-        }
-        return sync;
-    }
-
-    private Future invoke(Operation operation, int replicaIndex, int partitionId) {
-        final OperationService operationService = nodeEngine.getOperationService();
-        return operationService.createInvocationBuilder(InternalPartitionService.SERVICE_NAME, operation, partitionId)
-                .setTryCount(3)
-                .setTryPauseMillis(250)
-                .setReplicaIndex(replicaIndex)
-                .invoke();
-    }
-
-    private Operation createReplicaSyncStateOperation(long replicaVersion, int partitionId) {
-        final Operation op = new IsReplicaVersionSync(replicaVersion);
-        op.setService(this);
-        op.setNodeEngine(nodeEngine);
-        op.setOperationResponseHandler(createErrorLoggingResponseHandler(node.getLogger(IsReplicaVersionSync.class)));
-        op.setPartitionId(partitionId);
-
-        return op;
-    }
-
-    private boolean checkReplicaSyncState() {
-        if (!initialized) {
-            return true;
-        }
-
-        if (!hasMultipleMemberGroups()) {
-            return true;
-        }
-
-        final Address thisAddress = node.getThisAddress();
-        final Semaphore s = new Semaphore(0);
-        final AtomicBoolean ok = new AtomicBoolean(true);
-        final ExecutionCallback<Object> callback = new ExecutionCallback<Object>() {
-            @Override
-            public void onResponse(Object response) {
-                if (Boolean.FALSE.equals(response)) {
-                    ok.compareAndSet(true, false);
-                }
-                s.release();
-            }
-
-            @Override
-            public void onFailure(Throwable t) {
-                ok.compareAndSet(true, false);
-            }
-        };
-        int ownedCount = submitSyncReplicaOperations(thisAddress, s, ok, callback);
-        try {
-            if (ok.get()) {
-                int permits = ownedCount * getMaxAllowedBackupCount();
-                return s.tryAcquire(permits, REPLICA_SYNC_CHECK_TIMEOUT_SECONDS, TimeUnit.SECONDS) && ok.get();
-            } else {
-                return false;
-            }
-        } catch (InterruptedException ignored) {
-            return false;
-        }
-    }
-
-    private int submitSyncReplicaOperations(Address thisAddress, Semaphore s, AtomicBoolean ok,
-                                            ExecutionCallback callback) {
-
-        int ownedCount = 0;
-        ILogger responseLogger = node.getLogger(SyncReplicaVersion.class);
-        OperationResponseHandler responseHandler =
-                createErrorLoggingResponseHandler(responseLogger);
-
-        int maxBackupCount = getMaxAllowedBackupCount();
-
-        for (InternalPartitionImpl partition : partitions) {
-            Address owner = partition.getOwnerOrNull();
-            if (thisAddress.equals(owner)) {
-                for (int i = 1; i <= maxBackupCount; i++) {
-                    final Address replicaAddress = partition.getReplicaAddress(i);
-                    if (replicaAddress != null) {
-                        if (checkClusterStateForReplicaSync(replicaAddress)) {
-                            SyncReplicaVersion op = new SyncReplicaVersion(i, callback);
-                            op.setService(this);
-                            op.setNodeEngine(nodeEngine);
-                            op.setOperationResponseHandler(responseHandler);
-                            op.setPartitionId(partition.getPartitionId());
-                            nodeEngine.getOperationService().executeOperation(op);
-                        } else {
-                            s.release();
-                        }
-                    } else {
-                        ok.set(false);
-                        s.release();
-                    }
-                }
-                ownedCount++;
-            } else if (owner == null) {
-                ok.set(false);
-            }
-        }
-        return ownedCount;
-    }
-
-    private boolean checkClusterStateForReplicaSync(final Address address) {
-        final ClusterServiceImpl clusterService = node.clusterService;
-        final ClusterState clusterState = clusterService.getClusterState();
-
-        if (clusterState == ClusterState.ACTIVE || clusterState == ClusterState.IN_TRANSITION) {
-            return true;
-        }
-
-        return !clusterService.isMemberRemovedWhileClusterIsNotActive(address);
-    }
-
-    private boolean shouldWaitMigrationOrBackups(Level level) {
-        if (!preCheckShouldWaitMigrationOrBackups()) {
-            return false;
-        }
-
-        if (checkForActiveMigrations(level)) {
-            return true;
-        }
-
-        for (InternalPartitionImpl partition : partitions) {
-            if (partition.getReplicaAddress(1) == null) {
-                final boolean canTakeBackup = !isClusterFormedByOnlyLiteMembers();
-
-                if (canTakeBackup && logger.isLoggable(level)) {
-                    logger.log(level, "Should take backup of partitionId=" + partition.getPartitionId());
-                }
-
-                return canTakeBackup;
-            }
-        }
-        int replicaSyncProcesses = maxParallelReplications - replicaSyncProcessLock.availablePermits();
-        if (replicaSyncProcesses > 0) {
-            if (logger.isLoggable(level)) {
-                logger.log(level, "Processing replica sync requests: " + replicaSyncProcesses);
-            }
-            return true;
-        }
-        return false;
-    }
-
-    private boolean preCheckShouldWaitMigrationOrBackups() {
-        if (!initialized) {
-            return false;
-        }
-
-        return hasMultipleMemberGroups();
-    }
-
-    private boolean hasMultipleMemberGroups() {
-        return getMemberGroupsSize() >= 2;
-    }
-
-    private boolean checkForActiveMigrations(Level level) {
-        final MigrationInfo activeMigrationInfo = this.activeMigrationInfo;
-        if (activeMigrationInfo != null) {
-            if (logger.isLoggable(level)) {
-                logger.log(level, "Waiting for active migration: " + activeMigrationInfo);
-            }
-            return true;
-        }
-
-        int queueSize = migrationQueue.size();
-        if (queueSize != 0) {
-            if (logger.isLoggable(level)) {
-                logger.log(level, "Waiting for cluster migration tasks: " + queueSize);
-            }
-            return true;
-        }
-        return false;
+        return migrationManager.hasOnGoingMigration();
     }
 
     @Override
@@ -1586,18 +715,13 @@ public class InternalPartitionServiceImpl implements InternalPartitionService, M
     // Caution: Returning version array without copying for performance reasons. Callers must not modify this array!
     @Override
     public long[] incrementPartitionReplicaVersions(int partitionId, int backupCount) {
-        PartitionReplicaVersions replicaVersion = replicaVersions[partitionId];
-        return replicaVersion.incrementAndGet(backupCount);
+        return replicaManager.incrementPartitionReplicaVersions(partitionId, backupCount);
     }
 
     // called in operation threads
     @Override
     public void updatePartitionReplicaVersions(int partitionId, long[] versions, int replicaIndex) {
-        PartitionReplicaVersions partitionVersion = replicaVersions[partitionId];
-        if (!partitionVersion.update(versions, replicaIndex)) {
-            // this partition backup is behind the owner.
-            triggerPartitionReplicaSync(partitionId, replicaIndex, 0L);
-        }
+        replicaManager.updatePartitionReplicaVersions(partitionId, versions, replicaIndex);
     }
 
     @Override
@@ -1610,52 +734,18 @@ public class InternalPartitionServiceImpl implements InternalPartitionService, M
     // Caution: Returning version array without copying for performance reasons. Callers must not modify this array!
     @Override
     public long[] getPartitionReplicaVersions(int partitionId) {
-        return replicaVersions[partitionId].get();
+        return replicaManager.getPartitionReplicaVersions(partitionId);
     }
 
     // called in operation threads
     @Override
     public void setPartitionReplicaVersions(int partitionId, long[] versions, int replicaOffset) {
-        replicaVersions[partitionId].set(versions, replicaOffset);
+        replicaManager.setPartitionReplicaVersions(partitionId, versions, replicaOffset);
     }
 
     @Override
     public void clearPartitionReplicaVersions(int partitionId) {
-        replicaVersions[partitionId].clear();
-    }
-
-    // called in operation threads
-    public void finalizeReplicaSync(int partitionId, int replicaIndex, long[] versions) {
-        PartitionReplicaVersions replicaVersion = replicaVersions[partitionId];
-        replicaVersion.clear();
-        replicaVersion.set(versions, replicaIndex);
-        clearReplicaSyncRequest(partitionId, replicaIndex);
-    }
-
-    // called in operation threads
-    public void clearReplicaSyncRequest(int partitionId, int replicaIndex) {
-        ReplicaSyncInfo syncInfo = new ReplicaSyncInfo(partitionId, replicaIndex, null);
-        ReplicaSyncInfo currentSyncInfo = replicaSyncRequests.get(partitionId);
-
-        replicaSyncScheduler.cancelIfExists(partitionId, syncInfo);
-
-        if (syncInfo.equals(currentSyncInfo)
-                && replicaSyncRequests.compareAndSet(partitionId, currentSyncInfo, null)) {
-
-            releaseReplicaSyncPermit();
-        } else if (currentSyncInfo != null) {
-            if (logger.isFinestEnabled()) {
-                logger.finest("Not able to cancel sync! " + syncInfo + " VS Current " + currentSyncInfo);
-            }
-        }
-    }
-
-    public boolean tryToAcquireReplicaSyncPermit() {
-        return replicaSyncProcessLock.tryAcquire();
-    }
-
-    public void releaseReplicaSyncPermit() {
-        replicaSyncProcessLock.release();
+        replicaManager.clearPartitionReplicaVersions(partitionId);
     }
 
     @Override
@@ -1690,26 +780,11 @@ public class InternalPartitionServiceImpl implements InternalPartitionService, M
 
     @Override
     public void reset() {
-        migrationQueue.clear();
-        for (int k = 0; k < replicaSyncRequests.length(); k++) {
-            replicaSyncRequests.set(k, null);
-        }
-        replicaSyncScheduler.cancelAll();
-        // this is not sync with possibly running sync process
-        // permit count can exceed allowed parallelization count.
-        replicaSyncProcessLock.drainPermits();
-        replicaSyncProcessLock.release(maxParallelReplications);
-
         lock.lock();
         try {
-            initialized = false;
-            for (InternalPartitionImpl partition : partitions) {
-                partition.reset();
-            }
-            // TODO BASRI IS THIS SAFE?
-            activeMigrationInfo = null;
-            completedMigrations.clear();
-            stateVersion.set(0);
+            replicaManager.reset();
+            partitionStateManager.reset();
+            migrationManager.reset();
         } finally {
             lock.unlock();
         }
@@ -1717,20 +792,20 @@ public class InternalPartitionServiceImpl implements InternalPartitionService, M
 
     @Override
     public void pauseMigration() {
-        migrationAllowed.set(false);
+        migrationManager.pauseMigration();
     }
 
     @Override
     public void resumeMigration() {
-        migrationAllowed.set(true);
+        migrationManager.resumeMigration();
     }
 
     public boolean isReplicaSyncAllowed() {
-        return migrationAllowed.get();
+        return migrationManager.isMigrationAllowed();
     }
 
     public boolean isMigrationAllowed() {
-        if (migrationAllowed.get()) {
+        if (migrationManager.isMigrationAllowed()) {
             ClusterState clusterState = node.getClusterService().getClusterState();
             return clusterState == ClusterState.ACTIVE;
         }
@@ -1741,87 +816,43 @@ public class InternalPartitionServiceImpl implements InternalPartitionService, M
     @Override
     public void shutdown(boolean terminate) {
         logger.finest("Shutting down the partition service");
-        migrationThread.stopNow();
+        migrationManager.stop();
         reset();
     }
 
     @Override
     @Probe(name = "migrationQueueSize")
     public long getMigrationQueueSize() {
-        return migrationQueue.size();
+        return migrationManager.getMigrationQueueSize();
     }
 
     public PartitionServiceProxy getPartitionServiceProxy() {
         return proxy;
     }
 
-    private void sendMigrationEvent(final MigrationInfo migrationInfo, final MigrationStatus status) {
-        MemberImpl current = getMember(migrationInfo.getSource());
-        MemberImpl newOwner = getMember(migrationInfo.getDestination());
-        MigrationEvent event = new MigrationEvent(migrationInfo.getPartitionId(), current, newOwner, status);
-        EventService eventService = nodeEngine.getEventService();
-        Collection<EventRegistration> registrations = eventService.getRegistrations(SERVICE_NAME, MIGRATION_EVENT_TOPIC);
-        eventService.publishEvent(SERVICE_NAME, registrations, event, event.getPartitionId());
-    }
-
     @Override
     public String addMigrationListener(MigrationListener listener) {
-        if (listener == null) {
-            throw new NullPointerException("listener can't be null");
-        }
-
-        final MigrationListenerAdapter adapter = new MigrationListenerAdapter(listener);
-
-        EventService eventService = nodeEngine.getEventService();
-        EventRegistration registration = eventService.registerListener(SERVICE_NAME, MIGRATION_EVENT_TOPIC, adapter);
-        return registration.getId();
+        return partitionEventManager.addMigrationListener(listener);
     }
 
     @Override
     public boolean removeMigrationListener(String registrationId) {
-        if (registrationId == null) {
-            throw new NullPointerException("registrationId can't be null");
-        }
-
-        EventService eventService = nodeEngine.getEventService();
-        return eventService.deregisterListener(SERVICE_NAME, MIGRATION_EVENT_TOPIC, registrationId);
+        return partitionEventManager.removeMigrationListener(registrationId);
     }
 
     @Override
     public String addPartitionLostListener(PartitionLostListener listener) {
-        if (listener == null) {
-            throw new NullPointerException("listener can't be null");
-        }
-
-        final PartitionLostListenerAdapter adapter = new PartitionLostListenerAdapter(listener);
-
-        EventService eventService = nodeEngine.getEventService();
-        EventRegistration registration = eventService.registerListener(SERVICE_NAME, PARTITION_LOST_EVENT_TOPIC, adapter);
-        return registration.getId();
+        return partitionEventManager.addPartitionLostListener(listener);
     }
 
     @Override
     public String addLocalPartitionLostListener(PartitionLostListener listener) {
-        if (listener == null) {
-            throw new NullPointerException("listener can't be null");
-        }
-
-        final PartitionLostListenerAdapter adapter = new PartitionLostListenerAdapter(listener);
-
-        EventService eventService = nodeEngine.getEventService();
-        EventRegistration registration =
-                eventService.registerLocalListener(SERVICE_NAME, PARTITION_LOST_EVENT_TOPIC, adapter);
-        return registration.getId();
+        return partitionEventManager.addLocalPartitionLostListener(listener);
     }
 
     @Override
     public boolean removePartitionLostListener(String registrationId) {
-        if (registrationId == null) {
-            throw new NullPointerException("registrationId can't be null");
-        }
-
-        EventService eventService = nodeEngine.getEventService();
-        return eventService.deregisterListener(SERVICE_NAME, PARTITION_LOST_EVENT_TOPIC, registrationId);
+        return partitionEventManager.removePartitionLostListener(registrationId);
     }
 
     @Override
@@ -1832,597 +863,95 @@ public class InternalPartitionServiceImpl implements InternalPartitionService, M
     public void addPartitionListener(PartitionListener listener) {
         lock.lock();
         try {
-            PartitionListenerNode head = partitionListener.listenerHead;
-            partitionListener.listenerHead = new PartitionListenerNode(listener, head);
+            partitionListener.addChildListener(listener);
         } finally {
             lock.unlock();
         }
     }
 
     @Override
-    public String toString() {
-        return "PartitionManager[" + stateVersion + "] {\n\n" + "migrationQ: " + migrationQueue.size() + "\n}";
-    }
-
-    public Node getNode() {
-        return node;
-    }
-
-    @Override
     public boolean isPartitionOwner(int partitionId) {
-        InternalPartition partition = getPartition(partitionId);
-        return node.getThisAddress().equals(partition.getOwnerOrNull());
+        InternalPartition partition = partitionStateManager.getPartitionImpl(partitionId);
+        return partition.isLocal();
     }
 
     @Override
     public int getPartitionStateVersion() {
-        return stateVersion.get();
+        return partitionStateManager.getVersion();
     }
 
     @Override
     public void onPartitionLost(IPartitionLostEvent event) {
-        final PartitionLostEvent partitionLostEvent = new PartitionLostEvent(event.getPartitionId(), event.getLostReplicaIndex(),
-                event.getEventSource());
-        final EventService eventService = nodeEngine.getEventService();
-        final Collection<EventRegistration> registrations = eventService
-                .getRegistrations(SERVICE_NAME, PARTITION_LOST_EVENT_TOPIC);
-        eventService.publishEvent(SERVICE_NAME, registrations, partitionLostEvent, event.getPartitionId());
+        partitionEventManager.onPartitionLost(event);
     }
 
     public void setInternalMigrationListener(InternalMigrationListener listener) {
-        Preconditions.checkNotNull(listener);
-        internalMigrationListener = listener;
-    }
-
-    public void resetInternalMigrationListener() {
-        internalMigrationListener = new InternalMigrationListener.NopInternalMigrationListener();
+        migrationManager.setInternalMigrationListener(listener);
     }
 
     public InternalMigrationListener getInternalMigrationListener() {
-        return internalMigrationListener;
+        return migrationManager.getInternalMigrationListener();
+    }
+
+    public void resetInternalMigrationListener() {
+        migrationManager.resetInternalMigrationListener();
     }
 
     /**
      * @return copy of ongoing replica-sync operations
      */
     public List<ReplicaSyncInfo> getOngoingReplicaSyncRequests() {
-        final int length = replicaSyncRequests.length();
-        final List<ReplicaSyncInfo> replicaSyncRequestsList = new ArrayList<ReplicaSyncInfo>(length);
-        for (int i = 0; i < length; i++) {
-            final ReplicaSyncInfo replicaSyncInfo = replicaSyncRequests.get(i);
-            if (replicaSyncInfo != null) {
-                replicaSyncRequestsList.add(replicaSyncInfo);
-            }
-        }
-
-        return replicaSyncRequestsList;
+        return replicaManager.getOngoingReplicaSyncRequests();
     }
 
     /**
      * @return copy of scheduled replica-sync requests
      */
     public List<ScheduledEntry<Integer, ReplicaSyncInfo>> getScheduledReplicaSyncRequests() {
-        final List<ScheduledEntry<Integer, ReplicaSyncInfo>> entries = new ArrayList<ScheduledEntry<Integer, ReplicaSyncInfo>>();
-        for (int partitionId = 0; partitionId < partitionCount; partitionId++) {
-            final ScheduledEntry<Integer, ReplicaSyncInfo> entry = replicaSyncScheduler.get(partitionId);
-            if (entry != null) {
-                entries.add(entry);
-            }
-        }
-
-        return entries;
+        return replicaManager.getScheduledReplicaSyncRequests();
     }
 
-    private class SendPartitionRuntimeStateTask
-            implements Runnable {
-        @Override
-        public void run() {
-            if (node.isMaster() && node.getState() == NodeState.ACTIVE) {
-                if (migrationQueue.isNonEmpty() && isMigrationAllowed()) {
-                    logger.info("Remaining migration tasks in queue => " + migrationQueue.size());
-                }
-                publishPartitionRuntimeState();
-            }
-        }
+    public PartitionStateManager getPartitionStateManager() {
+        return partitionStateManager;
     }
 
-    private class SyncReplicaVersionTask implements Runnable {
-        @Override
-        public void run() {
-            if (node.nodeEngine.isRunning() && isReplicaSyncAllowed()) {
-                for (InternalPartitionImpl partition : partitions) {
-                    if (partition.isLocal()) {
-                        for (int index = 1; index < InternalPartition.MAX_REPLICA_COUNT; index++) {
-                            if (partition.getReplicaAddress(index) != null) {
-                                SyncReplicaVersion op = new SyncReplicaVersion(index, null);
-                                op.setService(InternalPartitionServiceImpl.this);
-                                op.setNodeEngine(nodeEngine);
-                                op.setOperationResponseHandler(
-                                        createErrorLoggingResponseHandler(node.getLogger(SyncReplicaVersion.class)));
-                                op.setPartitionId(partition.getPartitionId());
-                                nodeEngine.getOperationService().executeOperation(op);
-                            }
-                        }
-                    }
-                }
-            }
-        }
+    public MigrationManager getMigrationManager() {
+        return migrationManager;
     }
 
-    private class RepartitioningTask implements Runnable {
+    public PartitionReplicaManager getReplicaManager() {
+        return replicaManager;
+    }
+
+    public PartitionReplicaChecker getPartitionReplicaChecker() {
+        return partitionReplicaChecker;
+    }
+
+    public PartitionEventManager getPartitionEventManager() {
+        return partitionEventManager;
+    }
+
+    private class RepairPartitionTableTask implements Runnable {
+        private final Address deadAddress;
+
+        public RepairPartitionTableTask(Address deadAddress) {this.deadAddress = deadAddress;}
+
         @Override
         public void run() {
-            if (!node.isMaster()) {
+            if (!partitionStateManager.isInitialized()) {
                 return;
             }
 
-            lock.lock();
-            try {
-                if (!initialized) {
-                    return;
-                }
-                if (!isMigrationAllowed()) {
-                    return;
-                }
-
-                migrationQueue.clear();
-                PartitionStateGenerator psg = partitionStateGenerator;
-                Collection<MemberGroup> memberGroups = createMemberGroups();
-                Address[][] newState = psg.reArrange(memberGroups, partitions);
-
-                if (newState == null) {
-                    if (logger.isFinestEnabled()) {
-                        logger.finest("partition rearrangement couldn't be done. size of member groups: " + memberGroups.size());
-                    }
-                    return;
-                }
-
-                if (!isMigrationAllowed()) {
-                    return;
-                }
-
-                processNewPartitionState(newState);
-                syncPartitionRuntimeState();
-            } finally {
-                lock.unlock();
-            }
-        }
-
-        private void processNewPartitionState(Address[][] newState) {
-            int migrationCount = 0;
-            int lostCount = 0;
-            lastRepartitionTime.set(Clock.currentTimeMillis());
-            for (int partitionId = 0; partitionId < partitionCount; partitionId++) {
-                Address[] replicas = newState[partitionId];
-                InternalPartitionImpl currentPartition = partitions[partitionId];
-                Address currentOwner = currentPartition.getOwnerOrNull();
-                Address newOwner = replicas[0];
-
-                if (currentOwner == null) {
-                    // assign new owner for lost partition
-                    lostCount++;
-                    assignNewPartitionOwner(partitionId, replicas, currentPartition, newOwner);
-                } else if (newOwner != null && !currentOwner.equals(newOwner)) {
-                    if (logger.isFinestEnabled()) {
-                        logger.finest("PartitionToMigrate partitionId=" + partitionId
-                                + " replicas=" + Arrays.toString(replicas) + " currentOwner="
-                                + currentOwner + " newOwner=" + newOwner);
-                    }
-
-                    migrationCount++;
-                    migratePartitionToNewOwner(partitionId, replicas, currentOwner, newOwner);
-                } else {
-                    currentPartition.setReplicaAddresses(replicas);
-                }
-            }
-            logMigrationStatistics(migrationCount, lostCount);
-        }
-
-        private void logMigrationStatistics(int migrationCount, int lostCount) {
-            if (lostCount > 0) {
-                logger.warning("Assigning new owners for " + lostCount + " LOST partitions!");
+            Collection<MigrationInfo> migrationInfos = partitionStateManager.removeDeadAddress(deadAddress);
+            for (MigrationInfo migrationInfo : migrationInfos) {
+                // TODO: schedule migrations
             }
 
-            if (migrationCount > 0) {
-                logger.info("Re-partitioning cluster data... Migration queue size: " + migrationCount);
-            } else {
-                logger.info("Partition balance is ok, no need to re-partition cluster data... ");
-            }
-        }
-
-        private void migratePartitionToNewOwner(int partitionId, Address[] replicas, Address currentOwner, Address newOwner) {
-            MigrationInfo info = new MigrationInfo(partitionId, currentOwner, newOwner);
-            MigrateTask migrateTask = new MigrateTask(info, replicas);
-            migrationQueue.add(migrateTask);
-        }
-
-        private void assignNewPartitionOwner(int partitionId, Address[] replicas, InternalPartitionImpl currentPartition,
-                                             Address newOwner) {
-            currentPartition.setReplicaAddresses(replicas);
-            MigrationInfo migrationInfo = new MigrationInfo(partitionId, null, newOwner);
-            sendMigrationEvent(migrationInfo, MigrationStatus.STARTED);
-            sendMigrationEvent(migrationInfo, MigrationStatus.COMPLETED);
-        }
-
-        private boolean isMigrationAllowed() {
-            if (InternalPartitionServiceImpl.this.isMigrationAllowed()) {
-                return true;
-            }
-            migrationQueue.add(this);
-            return false;
+            syncPartitionRuntimeState();
         }
     }
 
-    class MigrateTask implements Runnable {
-        final MigrationInfo migrationInfo;
-        final Address[] addresses;
-
-        public MigrateTask(MigrationInfo migrationInfo, Address[] addresses) {
-            this.migrationInfo = migrationInfo;
-            this.addresses = addresses;
-            final MemberImpl masterMember = getMasterMember();
-            if (masterMember != null) {
-                migrationInfo.setMasterUuid(masterMember.getUuid());
-                migrationInfo.setMaster(masterMember.getAddress());
-            }
-        }
-
-        @Override
-        public void run() {
-            if (!node.isMaster()) {
-                return;
-            }
-            final MigrationRequestOperation migrationRequestOp = new MigrationRequestOperation(migrationInfo);
-            try {
-                MigrationInfo info = migrationInfo;
-                InternalPartitionImpl partition = partitions[info.getPartitionId()];
-                Address owner = partition.getOwnerOrNull();
-                if (owner == null) {
-                    logger.severe("ERROR: partition owner is not set! -> partitionId=" + info.getPartitionId()
-                            + " , " + partition + " -VS- " + info);
-                    return;
-                }
-                if (!owner.equals(info.getSource())) {
-                    logger.severe("ERROR: partition owner is not the source of migration! -> partitionId="
-                            + info.getPartitionId() + " , " + partition + " -VS- " + info + " found owner=" + owner);
-                    return;
-                }
-                internalMigrationListener.onMigrationStart(MigrationParticipant.MASTER, migrationInfo);
-                sendMigrationEvent(migrationInfo, MigrationStatus.STARTED);
-
-                Boolean result;
-                MemberImpl fromMember = getMember(migrationInfo.getSource());
-                if (logger.isFinestEnabled()) {
-                    logger.finest("Starting Migration: " + migrationInfo);
-                }
-                if (fromMember == null) {
-                    // Partition is lost! Assign new owner and exit.
-                    logger.warning("Partition is lost! Assign new owner and exit... partitionId=" + info.getPartitionId());
-                    result = Boolean.TRUE;
-                    // TODO BASRI UNDERSTAND HOW THIS PART WILL WORK. THIS CASE SHOULD CAN BE TESTED WITH OUR MIGRATION LISTENER
-                } else {
-                    result = executeMigrateOperation(migrationRequestOp, fromMember);
-                }
-                processMigrationResult(result);
-            } catch (Throwable t) {
-                final Level level = migrationEndpointsActive() ? Level.WARNING : Level.FINEST;
-                logger.log(level, "Error [" + t.getClass() + ": " + t.getMessage() + "] while executing " + migrationRequestOp);
-                logger.finest(t);
-                migrationOperationFailed();
-            }
-        }
-
-        private void processMigrationResult(Boolean result) {
-            if (Boolean.TRUE.equals(result)) {
-                if (logger.isFinestEnabled()) {
-                    logger.finest("Finished Migration: " + migrationInfo);
-                }
-                migrationOperationSucceeded();
-            } else {
-                final Level level = migrationEndpointsActive() ? Level.WARNING : Level.FINEST;
-                logger.log(level, "Migration failed: " + migrationInfo);
-                migrationOperationFailed();
-            }
-        }
-
-        private Boolean executeMigrateOperation(MigrationRequestOperation migrationRequestOp, MemberImpl fromMember) {
-            Future future = nodeEngine.getOperationService().createInvocationBuilder(SERVICE_NAME, migrationRequestOp,
-                    migrationInfo.getSource())
-                    .setCallTimeout(partitionMigrationTimeout)
-                    .setTryPauseMillis(DEFAULT_PAUSE_MILLIS).invoke();
-
-            try {
-                Object response = future.get();
-                return (Boolean) nodeEngine.toObject(response);
-            } catch (Throwable e) {
-                final Level level = nodeEngine.isRunning() && migrationEndpointsActive() ? Level.WARNING : Level.FINEST;
-                logger.log(level, "Failed migration from " + fromMember + " for " + migrationRequestOp.getMigrationInfo(), e);
-            }
-            return Boolean.FALSE;
-        }
-
-        private void migrationOperationFailed() {
-            internalMigrationListener.onMigrationComplete(MigrationParticipant.MASTER, migrationInfo, false);
-            lock.lock();
-            try {
-                addCompletedMigration(migrationInfo);
-                internalMigrationListener.onMigrationRollback(MigrationParticipant.MASTER, migrationInfo);
-                scheduleActiveMigrationFinalization(migrationInfo);
-                syncPartitionRuntimeState();
-            } finally {
-                lock.unlock();
-            }
-            sendMigrationEvent(migrationInfo, MigrationStatus.FAILED);
-
-            // Migration failed.
-            // Pause migration process for a small amount of time, if a migration attempt is failed.
-            // Otherwise, migration failures can do a busy spin until migration problem is resolved.
-            // Migration can fail either a node's just joined and not completed start yet or it's just left the cluster.
-            pauseMigration();
-            // Re-execute RepartitioningTask when all other migration tasks are done,
-            // an imbalance may occur because of this failure.
-            migrationQueue.add(new RepartitioningTask());
-            resumeMigrationEventually();
-        }
-
-        private void migrationOperationSucceeded() {
-            internalMigrationListener.onMigrationComplete(MigrationParticipant.MASTER, migrationInfo, true);
-            addCompletedMigration(migrationInfo);
-
-            boolean commitSuccessful = commitMigrationToDestination(migrationInfo, addresses);
-
-            lock.lock();
-            try {
-                if (commitSuccessful) {
-                    updatePartitionReplicaAddresses(migrationInfo.getPartitionId(), addresses);
-                    internalMigrationListener.onMigrationCommit(MigrationParticipant.MASTER, migrationInfo);
-                } else {
-                    internalMigrationListener.onMigrationRollback(MigrationParticipant.MASTER, migrationInfo);
-                }
-
-                scheduleActiveMigrationFinalization(migrationInfo);
-                syncPartitionRuntimeState();
-            } finally {
-                lock.unlock();
-            }
-            sendMigrationEvent(migrationInfo, MigrationStatus.COMPLETED);
-        }
-
-        private boolean migrationEndpointsActive() {
-            return getMember(migrationInfo.getSource()) != null
-                    && getMember(migrationInfo.getDestination()) != null;
-        }
-
-        @Override
-        public String toString() {
-            return getClass().getSimpleName() + "{" + "migrationInfo=" + migrationInfo + '}';
-        }
-    }
-
-    private class MigrationThread extends Thread implements Runnable {
-        private final long sleepTime = max(250L, partitionMigrationInterval);
-
-        MigrationThread(Node node) {
-            super(node.getHazelcastThreadGroup().getInternalThreadGroup(),
-                    node.getHazelcastThreadGroup().getThreadNamePrefix("migration"));
-        }
-
-        @Override
-        public void run() {
-            try {
-                while (!isInterrupted()) {
-                    doRun();
-                }
-            } catch (InterruptedException e) {
-                if (logger.isFinestEnabled()) {
-                    logger.finest("MigrationThread is interrupted: " + e.getMessage());
-                }
-            } catch (OutOfMemoryError e) {
-                OutOfMemoryErrorDispatcher.onOutOfMemory(e);
-            } finally {
-                migrationQueue.clear();
-            }
-        }
-
-        private void doRun() throws InterruptedException {
-            boolean migrating = false;
-            for (; ; ) {
-                if (!isMigrationAllowed()) {
-                    break;
-                }
-                Runnable r = migrationQueue.poll(1, TimeUnit.SECONDS);
-                if (r == null) {
-                    break;
-                }
-
-                migrating |= r instanceof MigrateTask;
-                processTask(r);
-                if (partitionMigrationInterval > 0) {
-                    Thread.sleep(partitionMigrationInterval);
-                }
-            }
-            boolean hasNoTasks = !migrationQueue.hasMigrationTasks();
-            if (hasNoTasks) {
-                if (migrating) {
-                    logger.info("All migration tasks have been completed, queues are empty.");
-                }
-                evictCompletedMigrations();
-                Thread.sleep(sleepTime);
-            } else if (!isMigrationAllowed()) {
-                Thread.sleep(sleepTime);
-            }
-        }
-
-        boolean processTask(Runnable r) {
-            try {
-                if (r == null || isInterrupted()) {
-                    return false;
-                }
-
-                r.run();
-            } catch (Throwable t) {
-                logger.warning(t);
-            } finally {
-                migrationQueue.afterTaskCompletion(r);
-            }
-
-            return true;
-        }
-
-        void stopNow() {
-            migrationQueue.clear();
-            interrupt();
-        }
-
-    }
-
-    private static final class InternalPartitionListener implements PartitionListener {
-        final Address thisAddress;
-        final InternalPartitionServiceImpl partitionService;
-        volatile PartitionListenerNode listenerHead;
-
-        private InternalPartitionListener(InternalPartitionServiceImpl partitionService, Address thisAddress) {
-            this.thisAddress = thisAddress;
-            this.partitionService = partitionService;
-        }
-
-        @Override
-        public void replicaChanged(PartitionReplicaChangeEvent event) {
-            final int partitionId = event.getPartitionId();
-            final int replicaIndex = event.getReplicaIndex();
-            final Address newAddress = event.getNewAddress();
-            final Address oldAddress = event.getOldAddress();
-            final PartitionReplicaChangeReason reason = event.getReason();
-
-            final boolean initialAssignment = event.getOldAddress() == null;
-
-            if (replicaIndex > 0) {
-                // backup replica owner changed!
-                if (thisAddress.equals(oldAddress)) {
-                    clearPartition(partitionId, replicaIndex);
-                } else if (thisAddress.equals(newAddress)) {
-                    synchronizePartition(partitionId, replicaIndex, reason, initialAssignment);
-                }
-            } else {
-                if (!initialAssignment && thisAddress.equals(newAddress)) {
-                    // it is possible that I might become owner while waiting for sync request from the previous owner.
-                    // I should check whether if have failed to get backups from the owner and lost the partition for
-                    // some backups.
-                    promoteFromBackups(partitionId, reason, oldAddress);
-                }
-                partitionService.cancelReplicaSync(partitionId);
-            }
-
-            Node node = partitionService.node;
-            if (replicaIndex == 0 && newAddress == null && node.isRunning() && node.joined()) {
-                logOwnerOfPartitionIsRemoved(event);
-            }
-            if (node.isMaster()) {
-                partitionService.stateVersion.incrementAndGet();
-            }
-
-            callListeners(event);
-        }
-
-        private void callListeners(PartitionReplicaChangeEvent event) {
-            PartitionListenerNode listenerNode = listenerHead;
-            while (listenerNode != null) {
-                try {
-                    listenerNode.listener.replicaChanged(event);
-                } catch (Throwable e) {
-                    partitionService.logger.warning("While calling PartitionListener: " + listenerNode.listener, e);
-                }
-                listenerNode = listenerNode.next;
-            }
-        }
-
-        private void clearPartition(final int partitionId, final int oldReplicaIndex) {
-            NodeEngine nodeEngine = partitionService.nodeEngine;
-            ClearReplicaOperation op = new ClearReplicaOperation(oldReplicaIndex);
-            op.setPartitionId(partitionId).setNodeEngine(nodeEngine).setService(partitionService);
-            nodeEngine.getOperationService().executeOperation(op);
-        }
-
-        private void synchronizePartition(int partitionId, int replicaIndex,
-                                          PartitionReplicaChangeReason reason, boolean initialAssignment) {
-            // if not initialized yet, no need to sync, since this is the initial partition assignment
-            if (partitionService.initialized) {
-                long delayMillis = 0L;
-                if (replicaIndex > 1) {
-                    // immediately trigger replica synchronization for the first backups
-                    // postpone replica synchronization for greater backups to a later time
-                    // high priority is 1st backups
-                    delayMillis = (long) (REPLICA_SYNC_RETRY_DELAY + (Math.random() * DEFAULT_REPLICA_SYNC_DELAY));
-                }
-
-                resetReplicaVersion(partitionId, replicaIndex, reason, initialAssignment);
-                partitionService.triggerPartitionReplicaSync(partitionId, replicaIndex, delayMillis);
-            }
-        }
-
-        private void resetReplicaVersion(int partitionId, int replicaIndex,
-                                         PartitionReplicaChangeReason reason, boolean initialAssignment) {
-            NodeEngine nodeEngine = partitionService.nodeEngine;
-            ResetReplicaVersionOperation op = new ResetReplicaVersionOperation(reason, initialAssignment);
-            op.setPartitionId(partitionId).setReplicaIndex(replicaIndex)
-                    .setNodeEngine(nodeEngine).setService(partitionService);
-            nodeEngine.getOperationService().executeOperation(op);
-        }
-
-        private void promoteFromBackups(int partitionId, PartitionReplicaChangeReason reason, Address oldAddress) {
-            NodeEngine nodeEngine = partitionService.nodeEngine;
-            PromoteFromBackupOperation op = new PromoteFromBackupOperation(reason, oldAddress);
-            op.setPartitionId(partitionId).setNodeEngine(nodeEngine).setService(partitionService);
-            nodeEngine.getOperationService().executeOperation(op);
-        }
-
-        private void logOwnerOfPartitionIsRemoved(PartitionReplicaChangeEvent event) {
-            String warning = "Owner of partition is being removed! "
-                    + "Possible data loss for partitionId=" + event.getPartitionId() + " , " + event;
-            partitionService.logger.warning(warning);
-        }
-    }
-
-    private static class ReplicaSyncEntryProcessor implements ScheduledEntryProcessor<Integer, ReplicaSyncInfo> {
-
-        final InternalPartitionServiceImpl partitionService;
-
-        ReplicaSyncEntryProcessor(InternalPartitionServiceImpl partitionService) {
-            this.partitionService = partitionService;
-        }
-
-        @Override
-        public void process(EntryTaskScheduler<Integer, ReplicaSyncInfo> scheduler,
-                            Collection<ScheduledEntry<Integer, ReplicaSyncInfo>> entries) {
-
-            for (ScheduledEntry<Integer, ReplicaSyncInfo> entry : entries) {
-                ReplicaSyncInfo syncInfo = entry.getValue();
-                int partitionId = syncInfo.partitionId;
-                if (partitionService.replicaSyncRequests.compareAndSet(partitionId, syncInfo, null)) {
-                    partitionService.releaseReplicaSyncPermit();
-                }
-
-                InternalPartitionImpl partition = partitionService.getPartitionImpl(partitionId);
-                int currentReplicaIndex = partition.getReplicaIndex(partitionService.node.getThisAddress());
-                if (currentReplicaIndex > 0) {
-                    partitionService.triggerPartitionReplicaSync(partitionId, currentReplicaIndex, 0L);
-                }
-            }
-        }
-    }
-
-    private static final class PartitionListenerNode {
-        final PartitionListener listener;
-        final PartitionListenerNode next;
-
-        PartitionListenerNode(PartitionListener listener, PartitionListenerNode next) {
-            this.listener = listener;
-            this.next = next;
-        }
-    }
-
-    // TODO: need a better task name!
-    private class FixPartitionTableTask implements Runnable {
+    private class FetchMostRecentPartitionTableTask implements Runnable {
         private final Address thisAddress = node.getThisAddress();
 
         public void run() {
@@ -2441,9 +970,11 @@ public class InternalPartitionServiceImpl implements InternalPartitionService, M
                 futures.add(future);
             }
 
-            int version = stateVersion.get();
+            int version = getPartitionStateVersion();
             PartitionRuntimeState newState = null;
 
+            // TODO BASRI HANDLE ACTIVE MIGRATION
+            // There might be 0, 1 (only for source -dest already committed-) or 2 (from source and dest) active migration objects
             Collection<MigrationInfo> activeMigrations = new ArrayList<MigrationInfo>();
             Collection<MigrationInfo> allCompletedMigrations = new HashSet<MigrationInfo>();
 
@@ -2459,6 +990,8 @@ public class InternalPartitionServiceImpl implements InternalPartitionService, M
                     if (state.getActiveMigration() != null) {
                         activeMigrations.add(state.getActiveMigration());
                     }
+                } catch (TargetNotMemberException e) {
+                    // ignore
                 } catch (MemberLeftException e) {
                     // ignore
                 } catch (InterruptedException e) {
@@ -2468,24 +1001,42 @@ public class InternalPartitionServiceImpl implements InternalPartitionService, M
                 }
             }
 
-            if (newState != null) {
-                applyNewState(newState, thisAddress);
-            }
-
             lock.lock();
             try {
-                // TODO: merge completed migrations
-                allCompletedMigrations.addAll(completedMigrations);
-                completedMigrations.clear();
-                completedMigrations.addAll(allCompletedMigrations);
+                if (migrationManager.getActiveMigration() != null) {
+                    activeMigrations.add(migrationManager.getActiveMigration());
+                }
 
-                // TODO: or should we ask all members about status of an ongoing migration?
-                // TODO: handle & rollback active migrations started but not completed yet
-//                        rollbackActiveMigrationsFromPreviousMaster(node.getLocalMember().getUuid());
+                // TODO: merge completed migrations
+                allCompletedMigrations.addAll(migrationManager.getCompletedMigrations());
+                migrationManager.setCompletedMigrations(allCompletedMigrations);
+
+                if (newState != null) {
+                    applyNewState(newState, thisAddress);
+                }
+
+                // TODO: iterate over active migrations and decide either to commit or to rollback
+                // we can (should ?) store status of active migration on both source and destination
+                // that way, we can see complete picture of a migration and decide whether we should commit
+                // or rollback.
+
+                // TODO: for now, just add activeMigrations into the completed migrations list
+                // eventually they'll be rolled back
+                for (MigrationInfo activeMigration : activeMigrations) {
+                    migrationManager.addCompletedMigration(activeMigration);
+                }
+                // TODO: remove local active migration?
+//                migrationManager.removeActiveMigration();
 
             } finally {
                 lock.unlock();
             }
         }
+    }
+
+    @Override
+    public String toString() {
+        return "PartitionManager[" + getPartitionStateVersion() + "] {\n\n"
+                + "migrationQ: " + getMigrationQueueSize() + "\n}";
     }
 }
