+++ hazelcast/src/main/java/com/hazelcast/internal/partition/operation/ClearReplicaOperation.java
@@ -19,6 +19,10 @@ package com.hazelcast.internal.partition.operation;
 import com.hazelcast.internal.partition.InternalPartition;
 import com.hazelcast.internal.partition.MigrationCycleOperation;
 import com.hazelcast.internal.partition.impl.InternalPartitionServiceImpl;
+import com.hazelcast.internal.partition.MigrationCycleOperation;
+import com.hazelcast.internal.partition.impl.InternalPartitionImpl;
+import com.hazelcast.internal.partition.impl.InternalPartitionServiceImpl;
+import com.hazelcast.internal.partition.impl.PartitionStateManager;
 import com.hazelcast.logging.ILogger;
 import com.hazelcast.nio.Address;
 import com.hazelcast.nio.ObjectDataInput;
@@ -45,7 +49,8 @@ public final class ClearReplicaOperation extends AbstractOperation
     public void run() throws Exception {
         int partitionId = getPartitionId();
         InternalPartitionServiceImpl partitionService = getService();
-        InternalPartition partition = partitionService.getPartitionImpl(partitionId);
+        PartitionStateManager partitionStateManager = partitionService.getPartitionStateManager();
+        InternalPartitionImpl partition = partitionStateManager.getPartitionImpl(partitionId);
         Address thisAddress = getNodeEngine().getThisAddress();
 
         // currentReplicaIndex == -1;               not owner/backup then clear data
@@ -54,7 +59,7 @@ public final class ClearReplicaOperation extends AbstractOperation
         int currentReplicaIndex = partition.getReplicaIndex(thisAddress);
         if (currentReplicaIndex == -1 || currentReplicaIndex > oldReplicaIndex) {
             if (currentReplicaIndex == -1) {
-                partitionService.cancelReplicaSync(partitionId);
+                partitionService.getReplicaManager().cancelReplicaSync(partitionId);
             }
             ILogger logger = getLogger();
             if (logger.isFinestEnabled()) {
@@ -66,7 +71,7 @@ public final class ClearReplicaOperation extends AbstractOperation
     }
 
     private void clearPartition(int partitionId, InternalPartitionServiceImpl partitionService) {
-        NodeEngineImpl nodeEngine = partitionService.getNode().getNodeEngine();
+        NodeEngineImpl nodeEngine = (NodeEngineImpl) getNodeEngine();
         final Collection<MigrationAwareService> services = nodeEngine.getServices(MigrationAwareService.class);
         for (MigrationAwareService service : services) {
             try {
