+++ hazelcast/src/main/java/com/hazelcast/internal/partition/operation/MigrationOperation.java
@@ -20,7 +20,9 @@ import com.hazelcast.core.HazelcastException;
 import com.hazelcast.internal.partition.InternalPartitionService;
 import com.hazelcast.internal.partition.MigrationInfo;
 import com.hazelcast.internal.partition.impl.InternalPartitionServiceImpl;
-import com.hazelcast.partition.impl.InternalMigrationListener.MigrationParticipant;
+import com.hazelcast.internal.partition.impl.InternalMigrationListener.MigrationParticipant;
+import com.hazelcast.internal.partition.impl.InternalPartitionServiceImpl;
+import com.hazelcast.internal.partition.impl.MigrationManager;
 import com.hazelcast.logging.ILogger;
 import com.hazelcast.nio.Address;
 import com.hazelcast.nio.ObjectDataInput;
@@ -169,7 +171,8 @@ public final class MigrationOperation extends BaseMigrationOperation {
 
     private boolean addActiveMigration() {
         InternalPartitionServiceImpl partitionService = getService();
-        return partitionService.addActiveMigration(migrationInfo);
+        MigrationManager migrationManager = partitionService.getMigrationManager();
+        return migrationManager.addActiveMigration(migrationInfo);
     }
 
     private void runMigrationTask(Operation op) throws Exception {
