+++ hazelcast/src/main/java/com/hazelcast/internal/partition/operation/ReplicaSyncRequest.java
@@ -16,6 +16,12 @@
 
 package com.hazelcast.internal.partition.operation;
 
+import com.hazelcast.internal.partition.impl.PartitionStateManager;
+import com.hazelcast.logging.ILogger;
+import com.hazelcast.nio.Address;
+import com.hazelcast.nio.ObjectDataInput;
+import com.hazelcast.nio.ObjectDataOutput;
+import com.hazelcast.internal.partition.MigrationCycleOperation;
 import com.hazelcast.internal.partition.InternalPartition;
 import com.hazelcast.internal.partition.InternalPartitionService;
 import com.hazelcast.internal.partition.MigrationCycleOperation;
@@ -89,13 +95,14 @@ public final class ReplicaSyncRequest extends Operation implements PartitionAwar
                 sendResponse(tasks);
             }
         } finally {
-            partitionService.releaseReplicaSyncPermit();
+            partitionService.getReplicaManager().releaseReplicaSyncPermit();
         }
     }
 
     private boolean preCheckReplicaSync(NodeEngineImpl nodeEngine, int partitionId, int replicaIndex) throws IOException {
         InternalPartitionServiceImpl partitionService = (InternalPartitionServiceImpl) nodeEngine.getPartitionService();
-        InternalPartitionImpl partition = partitionService.getPartitionImpl(partitionId);
+        PartitionStateManager partitionStateManager = partitionService.getPartitionStateManager();
+        InternalPartitionImpl partition = partitionStateManager.getPartitionImpl(partitionId);
         Address owner = partition.getOwnerOrNull();
         long[] replicaVersions = partitionService.getPartitionReplicaVersions(partitionId);
         long currentVersion = replicaVersions[replicaIndex - 1];
@@ -119,7 +126,7 @@ public final class ReplicaSyncRequest extends Operation implements PartitionAwar
             return false;
         }
 
-        if (!partitionService.tryToAcquireReplicaSyncPermit()) {
+        if (!partitionService.getReplicaManager().tryToAcquireReplicaSyncPermit()) {
             if (logger.isFinestEnabled()) {
                 logger.finest(
                         "Max parallel replication process limit exceeded! Could not run replica sync -> " + toString());
