+++ hazelcast/src/main/java/com/hazelcast/internal/partition/operation/ReplicaSyncResponse.java
@@ -20,6 +20,7 @@ import com.hazelcast.internal.partition.InternalPartitionService;
 import com.hazelcast.internal.partition.ReplicaErrorLogger;
 import com.hazelcast.internal.partition.impl.InternalPartitionImpl;
 import com.hazelcast.internal.partition.impl.InternalPartitionServiceImpl;
+import com.hazelcast.internal.partition.impl.PartitionStateManager;
 import com.hazelcast.logging.ILogger;
 import com.hazelcast.nio.Address;
 import com.hazelcast.nio.ObjectDataInput;
@@ -68,7 +69,8 @@ public class ReplicaSyncResponse extends Operation
         int partitionId = getPartitionId();
         int replicaIndex = getReplicaIndex();
 
-        InternalPartitionImpl partition = partitionService.getPartitionImpl(partitionId);
+        PartitionStateManager partitionStateManager = partitionService.getPartitionStateManager();
+        InternalPartitionImpl partition = partitionStateManager.getPartitionImpl(partitionId);
         Address thisAddress = nodeEngine.getThisAddress();
         int currentReplicaIndex = partition.getReplicaIndex(thisAddress);
         try {
@@ -90,13 +92,13 @@ public class ReplicaSyncResponse extends Operation
         int replicaIndex = getReplicaIndex();
 
         if (replicaIndex == currentReplicaIndex) {
-            partitionService.finalizeReplicaSync(partitionId, replicaIndex, replicaVersions);
+            partitionService.getReplicaManager().finalizeReplicaSync(partitionId, replicaIndex, replicaVersions);
         } else {
-            partitionService.clearReplicaSyncRequest(partitionId, replicaIndex);
+            partitionService.getReplicaManager().clearReplicaSyncRequest(partitionId, replicaIndex);
             if (currentReplicaIndex < 0) {
                 partitionService.clearPartitionReplicaVersions(partitionId);
             } else if (currentReplicaIndex > 0) {
-                partitionService.triggerPartitionReplicaSync(partitionId, currentReplicaIndex, 0);
+                partitionService.getReplicaManager().triggerPartitionReplicaSync(partitionId, currentReplicaIndex, 0);
             }
         }
     }
