+++ hazelcast/src/test/java/com/hazelcast/partition/impl/MigrationCommitTest.java
@@ -5,10 +5,11 @@ import com.hazelcast.config.ListenerConfig;
 import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.core.HazelcastInstanceAware;
 import com.hazelcast.instance.GroupProperty;
+import com.hazelcast.internal.partition.InternalPartition;
+import com.hazelcast.internal.partition.InternalPartitionService;
+import com.hazelcast.internal.partition.MigrationInfo;
+import com.hazelcast.internal.partition.impl.InternalMigrationListener;
 import com.hazelcast.nio.Address;
-import com.hazelcast.partition.InternalPartition;
-import com.hazelcast.partition.InternalPartitionService;
-import com.hazelcast.partition.MigrationInfo;
 import com.hazelcast.partition.impl.InternalMigrationListenerTest.InternalMigrationListenerImpl;
 import com.hazelcast.partition.impl.InternalMigrationListenerTest.MigrationProgressNotification;
 import com.hazelcast.test.HazelcastParallelClassRunner;
